% Generated by roxygen2 (4.0.1): do not edit by hand
\name{abind3}
\alias{abind3}
\title{Combine two-dimensional arrays}
\usage{
abind3(...)
}
\arguments{
\item{...}{any number of two dimensional objects, to be bound together along
the third dimension.}
}
\description{
Binds two dimensional arrays along the third dimension.
}

