% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{addAlpha}
\alias{addAlpha}
\title{Adds alpha transparency to any color}
\usage{
addAlpha(col, alpha)
}
\arguments{
\item{col}{color}

\item{alpha}{alpha transparency, should be between 0 and 1}
}
\value{
the hexidecimal value of the color with the added alpha channel
}
\description{
Adds alpha transparency to any color
}

