% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addGradientLegend}
\alias{addGradientLegend}
\title{Plot a color palette legend}
\usage{
addGradientLegend(palette, palette.vlim, legend.vlim, horizontal, main, xlim,
  ylim)
}
\arguments{
\item{palette}{color palette.}

\item{palette.vlim}{limits of the values mapping to the extremities of the
color palette.}

\item{legend.vlim}{limits of the values to display on the legend}

\item{horizontal}{logical; if \code{TRUE} the legend is plotted horizontally,
otherwise vertically.}

\item{main}{title of the legend.}

\item{xlim}{xlim relative to the plotting region of the rest of the plot.}

\item{ylim}{ylim relative to the plotting region of the rest of the plot.}
}
\description{
Add a legend to a plot window.
}

