% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-defs.R
\name{adj_param}
\alias{adj_param}
\title{Template parameters}
\arguments{
\item{adj}{a \code{\link[=bigMatrix-class]{bigMatrix}} containing the
pairwise gene adjacencies.}
}
\description{
Template parameters to be imported into other function documentation. This 
is not intended to be a stand-alone help file.
}

