% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{adjacency.palette}
\alias{adjacency.palette}
\title{Color palette for adjacency heatmaps}
\usage{
adjacency.palette()
}
\description{
RColorBrewer palette "RdYlBu" with the middle color replaced with white.
}

