% Generated by roxygen2 (4.0.1): do not edit by hand
\name{AdjProps}
\alias{AdjProps}
\title{Calculate Network Properties}
\usage{
AdjProps(pAdjacency, subsetIndices)
}
\arguments{
\item{pAdjacency}{SEXP container for the pointer to the adjacency matrix}

\item{subsetIndices}{indices of the subset of the network to calculate
  the mean adjacency for.}
}
\value{
A List containing:
  \enumerate{
    \item{\emph{kIM}:}{The weighted within-subset degree for each node.}
    \item{\emph{MAR}:}{The maximum adjacency ratio for each node.}
    \item{\emph{meanAdj}:}{The mean absolute edge weight of the network subset.}
    \item{\emph{meanKIM}:}{The mean within-subset degree.}
    \item{\emph{meanMAR}:}{The mean maximum adjacency ratio.}
  }
}
\description{
Calculate Network Properties
}

