% Generated by roxygen2 (4.0.1): do not edit by hand
\name{AllFinite}
\alias{AllFinite}
\title{Check the elements of a `big.matrix`}
\usage{
AllFinite(pDat)
}
\arguments{
\item{pDat}{SEXP container for the pointer to the
\code{\link[bigmemory]{big.matrix}} to be checked.}
}
\value{
\code{TRUE} if all values are finite, \code{FALSE} otherwise
}
\description{
Are all the values finite?
}

