% Generated by roxygen2 (4.0.1): do not edit by hand
\name{allFinite}
\alias{allFinite}
\title{Check if all entries of a `big.matrix` are Finite}
\usage{
allFinite(x)
}
\arguments{
\item{x}{a \code{\link[bigmemory]{big.matrix}}}
}
\value{
\code{FALSE} if \code{x} contains any of: \code{NA}, \code{NaN}, \code{Inf},
 of \code{-Inf}, and \code{TRUE} otherwise.
}
\description{
Check if all entries of a `big.matrix` are Finite
}

