% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{any.big.matrix}
\alias{any.big.matrix}
\title{Check if any objects are a 'big.matrix'}
\usage{
\method{any}{big.matrix}(...)
}
\arguments{
\item{...}{objects to check.}
}
\value{
\code{TRUE} if the class of any object in the list of input 
 arguments is a "big.matrix".
}
\description{
Check if any objects are a 'big.matrix'
}

