% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{any.disk.matrix}
\alias{any.disk.matrix}
\title{Check if any objects are a 'disk.matrix'}
\usage{
\method{any}{disk.matrix}(...)
}
\arguments{
\item{...}{objects to check.}
}
\value{
\code{TRUE} if the class of any object in the list of input 
 arguments is a "disk.matrix".
}
\description{
Check if any objects are a 'disk.matrix'
}
\keyword{internal}

