% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{areColors}
\alias{areColors}
\title{Check if a character vector of colors is valid}
\usage{
areColors(colvec)
}
\arguments{
\item{colvec}{a character vectors of colors (hex or name) to validate.}
}
\description{
Courtesy of Josh O'Brien's stackoverflow answer at
\url{http://stackoverflow.com/a/13290832/2341679}
}
\keyword{internal}

