% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{bigMatrix-class}
\alias{[,bigMatrix,ANY,ANY,ANY-method}
\alias{[,bigMatrix,ANY,missing,ANY-method}
\alias{[,bigMatrix,missing,ANY,ANY-method}
\alias{[,bigMatrix,missing,missing,ANY-method}
\alias{[<-,bigMatrix,ANY,ANY-method}
\alias{[<-,bigMatrix,ANY,missing-method}
\alias{[<-,bigMatrix,missing,ANY-method}
\alias{[<-,bigMatrix,missing,missing-method}
\alias{bigMatrix-class}
\alias{dim,bigMatrix-method}
\alias{dimnames,bigMatrix-method}
\alias{dimnames<-,bigMatrix,ANY-method}
\title{The bigMatrix class}
\description{
A \code{bigMatrix} is an object which points to a data matrix stored on disk,
which is accessible via shared memory. This has a number of advantages,
including instananeous load times and shared access to the same object across
multiple R sessions, enabling massively parallel operations on a matrix
without the burden of excessive memory consumption.
}
\details{
A \code{bigMatrix} object is simply a wrapper around a
\code{\link[bigmemory]{big.matrix}} object from the
\code{\link[bigmemory]{bigmemory}} package.

The advantage of the \code{bigMatrix} over \code{big.matrix} is that it keeps
the object in a "detached" state: the \code{big.matrix} is only attached on
access through some other method (e.g. subsetting or element access). This
allows for finer control over memory usage in the
\code{\link{modulePreservation}} routine for users with limited RAM who wish
to run \code{modulePreservation} in the background on their laptop or
desktop. This is also nice for RStudio users, as it prevents R from crashing
when reloading a session containing the \code{bigMatrix} objects.

The authors of the \code{bigmemory} package also explicitly warn against
storage of row and column names in a \code{big.matrix} as it can slow down
computation. A \code{bigMatrix} object handles these by keeping the dimension
names stored in R, which offers (minor) speed improvements when running
\code{modulePreservation} in comparison to keeping to row and column names
stored in shared memory data structure.
}
\section{Slots}{

\describe{
\item{\code{descriptor}}{path of the descriptor file for the big.matrix.}

\item{\code{matrix}}{either the big.matrix object, or empty, depending on the value
of \code{atttached}.}

\item{\code{rownames}}{Optional vector of rownames.}

\item{\code{colnames}}{Optional vector of colnames.}

\item{\code{attached}}{Logical; \code{TRUE} when the big.matrix is attached in the
 R session, and \code{FALSE} otherwise.}
}}
\seealso{
\code{\link[bigmemory]{big.matrix}},
 \code{\link{bigMatrix-get}}
 \code{\link{bigMatrix-out}}
}

