% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bigMatrix-get}
\alias{as.bigMatrix}
\alias{bigMatrix-get}
\alias{load.bigMatrix}
\alias{read.bigMatrix}
\alias{save.as.bigMatrix}
\title{Loading in data as a 'bigMatrix'}
\usage{
save.as.bigMatrix(x, backingfile,
  type = options("bigmemory.default.type")[[1]])

load.bigMatrix(backingfile)

as.bigMatrix(x, backingfile, type = options("bigmemory.default.type")[[1]])

read.bigMatrix(file, backingfile,
  type = options("bigmemory.default.type")[[1]], row.names = TRUE,
  header = TRUE, ...)
}
\arguments{
\item{x}{a regular matrix to convert}

\item{backingfile}{location on disk the bigMatrix is, or will be, stored at.}

\item{type}{the type of the atomic element
(\code{options()$bigmemory.default.type} by default - "\code{double}" -
but can be changed by the user to "\code{integer}", "\code{short}", or
"\code{char}").}

\item{file}{the name of the file which the (non-bigMatrix) data are to be
read from (see \code{\link[bigmemory]{read.big.matrix}}).}

\item{row.names}{logical; does the first column of the file to be read in
contain row names?}

\item{header}{logical; does the first line of the file to be read in contain
column names?}

\item{...}{additional arguments to pass to
 \code{\link[bigmemory]{read.big.matrix}}.}
}
\value{
\code{read.bigMatrix}, \code{load.bigMatrix}, \code{as.bigMatrix}: an
  object of class \code{\link[=bigMatrix-class]{bigMatrix}}.
}
\description{
Functions for converting to, reading in, and loading a
\code{\link[=bigMatrix-class]{bigMatrix}} object.
}
\details{
\code{\link[=bigMatrix-class]{bigMatrix}} objects are stored on disk using
across four different files per object. The \code{backingname} designates the
basename of these files, and the backingpath designates the directory they
are stored in. The matrix data are stored in the file with the ".bin"
extension. The file with the ".desc" extension contains a description
allowing R to load in this data (see
\code{\link[bigmemory]{attach.big.matrix}}). Row and column names are stored
separately in the files ending with "_rownames.txt" and "_colnames.txt".

\code{save.as.bigMatrix} takes a standard matrix and writes it out as a
\code{bigMatrix}. This can then be loaded into an R session using
\code{load.bigMatrix}. \code{as.bigMatrix} combines these two steps, taking
a standard matrix and returning a \code{bigMatrix} object. Finally,
\code{read.bigMatrix} can be used to read in matrix data from a file (i.e.
data recongisiable by \code{read.table}) as a \code{bigMatrix} object.

\code{load.bigMatrix} can also be used to load in a
\code{\link[bigmemory]{big.matrix}} as a 'bigMarix', but existing row and
column information will be stripped out from the existing backing file. This
means if you later load in the data as a 'big.matrix' instead of a
'bigMatrix' the row and column names will be missing unless you convert back
using \code{\link{as.big.matrix}}.
}
\seealso{
\code{\link[=bigMatrix-class]{bigMatrix}},
 \code{\link[bigmemory]{big.matrix}},
 \code{\link[bigmemory]{read.big.matrix}}
}

