% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bigMatrix-io}
\alias{as.bigMatrix}
\alias{bigMatrix-io}
\alias{load.bigMatrix}
\alias{read.bigMatrix}
\alias{write.bigMatrix}
\title{bigMatrix input and output}
\usage{
write.bigMatrix(x, backingname, backingpath,
  type = options("bigmemory.default.type"))

load.bigMatrix(backingname, backingpath)

as.bigMatrix(x, backingname, backingpath,
  type = options("bigmemory.default.type"))

read.bigMatrix(file, backingname, backingpath,
  type = options("bigmemory.default.type"), ...)
}
\arguments{
\item{x}{matrix to write out as a file-backed `big.matrix`.}

\item{backingname}{basename (filename without an extension) the
bigMatrix object is, or will be, stored in.}

\item{backingpath}{directory the bigMatrix is, or will be, stored in.}

\item{file}{the name of the file which the (non-bigMatrix) data are to be
read from using \code{\link[utils]{read.table}}.}

\item{...}{additional arguments to pass to \code{read.table}}

\item{the}{type of the atomic element
(\code{options()$bigmemory.default.type} by default – "\code{double}" –
but can be changed by the user to "\code{integer}", "\code{short}", or
"\code{char}").}
}
\value{
\code{read.bigMatrix}, \code{load.bigMatrix}, \code{as.bigMatrix}: an
  object of class \code{\link{bigMatrix}}.
}
\description{
Functions for converting to, reading in, and loading a
\code{\link{bigMatrix}} object.
}
\details{
\code{\link{bigMatrix}} objects are stored on disk using across four
different files per object. The \code{backingname} designates the basename
of these files, and the backingpath designates the directory they are stored
in. The matrix data are stored in the file with the ".bin" extension. The
file with the ".desc" extension contains a description allowing R to load in
this data (see \code{\link[bigmemory]{attach.big.matrix}}). Row and
column names are stored separately in the files ending with "_rownames.txt"
and "_colnames.txt".

\code{write.bigMatrix} takes a standard matrix and writes it out as a
\code{bigMatrix}. This can then be loaded into an R session using
\code{load.bigMatrix}. \code{as.bigMatrix} combines these two steps, taking
a standard matrix and returning a \code{bigMatrix} object. Finally,
\code{read.bigMatrix} can be used to read in matrix data from a file (i.e.
using \code{read.table}) as a \code{bigMatrix} object (note that the
appropriate backing files will also be created on disk).
}
\seealso{
\code{\link{bigMatrix}},
 \code{\link[bigmemory]{big.matrix}},
 \code{\link[utils]{read.table}}
}

