% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigMatrix-methods.R
\docType{methods}
\name{bigMatrix-methods}
\alias{[,bigMatrix,ANY,ANY,ANY-method}
\alias{[,bigMatrix,ANY,missing,ANY-method}
\alias{[,bigMatrix,missing,ANY,ANY-method}
\alias{[,bigMatrix,missing,missing,ANY-method}
\alias{[<-,bigMatrix,ANY,ANY-method}
\alias{[<-,bigMatrix,ANY,missing-method}
\alias{[<-,bigMatrix,missing,ANY-method}
\alias{[<-,bigMatrix,missing,missing-method}
\alias{as.big.matrix,bigMatrix-method}
\alias{as.matrix,bigMatrix-method}
\alias{bigMatrix-methods}
\alias{dim,bigMatrix-method}
\alias{dimnames,bigMatrix-method}
\alias{dimnames<-,bigMatrix,ANY-method}
\alias{head,bigMatrix-method}
\alias{ncol,bigMatrix-method}
\alias{nrow,bigMatrix-method}
\alias{print,bigMatrix-method}
\alias{show,bigMatrix-method}
\alias{tail,bigMatrix-method}
\alias{typeof,bigMatrix-method}
\title{Methods for interacting with and accessing 'bigMatrix' objects}
\usage{
\S4method{as.matrix}{bigMatrix}(x)

\S4method{as.big.matrix}{bigMatrix}(x)

\S4method{show}{bigMatrix}(object)

\S4method{print}{bigMatrix}(x)

\S4method{dim}{bigMatrix}(x)

\S4method{nrow}{bigMatrix}(x)

\S4method{ncol}{bigMatrix}(x)

\S4method{typeof}{bigMatrix}(x)

\S4method{head}{bigMatrix}(x, n = 6)

\S4method{tail}{bigMatrix}(x, n = 6)

\S4method{dimnames}{bigMatrix}(x)

\S4method{dimnames}{bigMatrix,ANY}(x) <- value

\S4method{[}{bigMatrix,ANY,ANY,ANY}(x, i, j, drop)

\S4method{[}{bigMatrix,ANY,missing,ANY}(x, i, j, drop)

\S4method{[}{bigMatrix,missing,ANY,ANY}(x, i, j, drop)

\S4method{[}{bigMatrix,missing,missing,ANY}(x, i, j, drop)

\S4method{[}{bigMatrix,ANY,ANY}(x, i, j) <- value

\S4method{[}{bigMatrix,ANY,missing}(x, i, j) <- value

\S4method{[}{bigMatrix,missing,ANY}(x, i, j) <- value

\S4method{[}{bigMatrix,missing,missing}(x, i, j) <- value
}
\arguments{
\item{x, object}{a \code{bigMatrix} object.}

\item{n}{number of rows to show in \code{head} or \code{tail}}

\item{value}{a possivle value for \code{dimnames(x)}. See 
\link[=base]{dimnames}.}

\item{i, j}{indices specifying elements to extract. Indices can be 
\code{numeric}, \code{character} vectors, \code{logical} vectors, or empty
(missing). See \link[=base]{Extract}.}

\item{drop}{logical; If \code{TRUE} the result is coerced to a vector when
extracting a single row, column, or element.}
}
\value{
\code{as.matrix}: a \code{matrix}.

\code{as.big.matrix}: a \code{\link[bigmemory]{big.matrix}} object.
}
\description{
Standard methods for accessing \code{\link{bigMatrix}} objects.
}
\details{
\code{bigMatrix} objects must be converted back to \code{matrix} objects to
perform matrix algebra, e.g. transposing (\code{\link[base]{t}}), either 
through \code{as.matrix} or by extracting elements of interest through 
\code{[}.
}
\examples{
\dontrun{
# load in example data, correlation, and network matrices for a discovery and test dataset:
data("NetRep")

# Convert them to the 'bigMatrix' format:
discovery_data <- as.bigMatrix(discovery_data)
discovery_correlation <- as.bigMatrix(discovery_correlation)
discovery_network <- as.bigMatrix(discovery_network)
test_data <- as.bigMatrix(test_data)
test_correlation <- as.bigMatrix(test_correlation)
test_network <- as.bigMatrix(test_network)

# 'bigMatrix' objects can be manipulated the same way as regular matrices:
head(discovery_data)
test_network[1:5, 1:5]
discovery_data[,"Node_1"]
dim(test_data)
rownames(test_data)
nrow(test_network)
ncol(test_correlation)
is.bigMatrix(discovery_data)
typeof(discovery_data)

# For matrix algebra the whole matrix must be copied into memory first:
discovery_data <- as.matrix(discovery_data)
test_data <- test_data[,] # equivalent to 'as.matrix'
t(test_network[,])
}
}
\seealso{
\code{\link{bigMatrix}}
}

