% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigMatrix-io.R
\docType{methods}
\name{bigMatrix-out}
\alias{as.big.matrix,bigMatrix-method}
\alias{as.matrix,bigMatrix-method}
\alias{bigMatrix-out}
\alias{write.bigMatrix}
\title{Writing out data, or converting from, a 'bigMatrix'.}
\usage{
write.bigMatrix(x, file, ...)

\S4method{as.matrix}{bigMatrix}(x)

\S4method{as.big.matrix}{bigMatrix}(x)
}
\arguments{
\item{x}{an object of type 'bigMatrix'.}

\item{file}{file to write out the data from \code{x} into.}

\item{...}{additional arguments to pass to 
\code{\link[bigmemory]{write.big.matrix}}.}
}
\description{
Functions for converting 'bigMatrix' objects to other classes.
}
\details{
\code{as.big.matrix} converts to a \code{\link[bigmemory]{big.matrix}}
object.
\code{as.matrix} converts to a \code{\link[base]{matrix}} object.
\code{write.bigMatrix} will write out the data into a regularly structure
file (see \code{\link[utils]{write.table}})
}

