% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bigMatrix-state}
\alias{attach.bigMatrix}
\alias{bigMatrix-state}
\alias{detach.bigMatrix}
\title{Attaching and detaching a bigMatrix}
\usage{
attach.bigMatrix(x)

detach.bigMatrix(x)
}
\arguments{
\item{x}{the bigMatrix object to attach or detach}
}
\value{
The bigMatrix object with its state updated.
}
\description{
In the attached state, R is aware of the \code{\link[bigmemory]{big.matrix}}
pointer object. Repeated access of different parts of the matrix, i.e.
through random sampling for permutation analysis will eventually load the
entire matrix into RAM. In the detached state, R is only aware of the
location of the 'big.matrix' descriptor file, and the big.matrix is attached
and removed every time its elements are requested.
}

