% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigMatrix.R
\name{load.bigMatrix}
\alias{load.bigMatrix}
\title{Load a 'bigMatrix' (deprecated)}
\usage{
load.bigMatrix(backingfile)
}
\arguments{
\item{backingfile}{path to the backingfile for the \code{'bigMatrix'}. The
file extension must be omitted.}
}
\description{
The \code{'bigMatrix'} class is no longer implemented in the \code{NetRep}
package: the shared memory approach was incompatabile with high performance
compute clusters, so the parallel permutation procedure has been translated
into C++ code (which is also much faster). For those who had saved their 
data in the \code{'bigMatrix'} format, this function will convert these 
objects to the similar \code{'\link[bigmemory]{big.matrix}'} class
from the \pkg{bigmemory} package.
}
\details{
A note for users using multi-node high performance clusters:
  \code{'big.matrix'} objects are not suitable for general usage. Access
  to file-backed shared memory segments on multi-node systems is very slow
  due to consistency checks performed by the operating system. We address
  this limitation by copying the data from \code{'big.matrix'} objects into
  new \code{'matrix'} objects. This can take several minutes, depending on
  the size of the total network.
}

