% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calcReplStats}
\alias{calcReplStats}
\title{Calculate between-dataset network statistics.}
\usage{
calcReplStats(discDat, discAdj, testDat, testAdj, nodeDict)
}
\arguments{
\item{discDat}{Underlying data for the reference network.}

\item{discAdj}{Adjacency matrix for the discovery network.}

\item{testDat}{Underlying data for the test network.}

\item{testAdj}{Adjacency matrix for the test network.}

\item{nodeDict}{A list with four elements, which map the nodes of the network
subset of interest to their indices in the \code{discDat}, \code{discAdj},
\code{testDat}, and \code{testAdj} respectively.}
}
\value{
A vector of preservation scores for the network subset.
}
\description{
For a given network subset, as defined by the nodes in the \code{nodeDict},
calculate
}
\references{
Langfelder, Peter; Luo, Rui; Oldham, Michael C.; and Horvath,
   Steve. Is My Network Module Preserved and Reproducible?. PLoS
   Computational Biology, 2011.
}

