% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calcSharedTestStats}
\alias{calcSharedTestStats}
\title{Network statistics from combined data}
\usage{
calcSharedTestStats(discAdj, discIndices, testAdj, testIndices)
}
\arguments{
\item{discAdj,testAdj}{\code{\link[bigmemory]{big.matrix}} objects for the
\emph{discovery} and \emph{test} networks.}

\item{discIndices,testIndices}{indices of the network subset in the
\emph{discovery} and \emph{test} networks.}
}
\value{
A vector of test statistics.
}
\description{
For some statistics it does not make sense to calculate the necessary
components in advance due to large memory overhead, or logic that doesn't
separate nicely. This function deals with those statistics.
}
\references{
Langfelder, P., Luo, R., Oldham, M. C. & Horvath, S. \emph{Is my network
module preserved and reproducible?} PLoS Comput. Biol. \strong{7}, e1001057
(2011).
}
\seealso{
\code{\link[=subsetProps]{Network subset topology}}
  \code{\link{netRep}}
}

