% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calcSplitTestStats}
\alias{calcSplitTestStats}
\title{Calculate network statistics from pre-calculated network properties}
\usage{
calcSplitTestStats(discProps, testProps)
}
\arguments{
\item{discProps,testProps}{properties of the network subset in the each of
the \emph{discovery} and \emph{test} datasets, respectively, as calculated
by \code{\link{subsetProps}}.}
}
\value{
A vector of test statistics.
}
\description{
Calculate network statistics from pre-calculated network properties
}
\details{
All network statistics can be fundamentally broken into the calculation of
some properties in the discovery network, and some calculation of properties
in the test network. It makes sense to split these calculations, so that we
only have to calculate the properties in the discovery network/data once.

However, for some statistics it does not make sense to do this due to memory
overhead (e.g. \emph{corAdj} would require storing a second copy of most of
the discovery network in vector form). For these statistics, see
\code{\link{calcSharedTestStats}}.
}
\section{Background}{

The returned test statistics indicate the replication/preservation of each
network property for a network subset in the two datasets provided in the
\code{discProps} and \code{testProps} arguments.

Evaluating the significance of these test statistics requires generation of a
null distribution for each, by randomly sampling network subsets of the same
size in the \emph{test} dataset, calculating their replication, and
evaluating how extreme the observed test statistic is, in comparison to test
statistics drawn from random permutation. This procedure is performed by the
main function of this package, \code{\link{netRep}}.
}
\references{
Langfelder, P., Luo, R., Oldham, M. C. & Horvath, S. \emph{Is my network
module preserved and reproducible?} PLoS Comput. Biol. \strong{7}, e1001057
(2011).
}
\seealso{
\code{\link[=subsetProps]{Network subset topology}}
  \code{\link{netRep}}
}

