% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-stats.R
\name{calcStats}
\alias{calcStats}
\title{Calculate module preservation statistics.}
\usage{
calcStats(discProps, testProps, discCor, discIndices, testCor, testIndices)
}
\arguments{
\item{discProps}{module properties in the \emph{discovery} dataset calculated 
using \code{\link{netProps}} and \code{\link{dataProps}}.}

\item{testProps}{module properties in the \emph{test} dataset calculated 
using \code{\link{netProps}} and \code{\link{dataProps}}.}

\item{discCor}{'bigMatrix' containing the correlation structure matrix 
for the \emph{discovery} dataset.}

\item{discIndices}{module indices in the discovery dataset.}

\item{testCor}{'bigMatrix' containing the correlation structure matrix 
for the \emph{test} dataset.}

\item{testIndices}{module indices in the test dataset.}
}
\value{
a vector containing the values for all the module preservation
statistics: \emph{avg.weight}, \emph{cor.degree}, 
\emph{coherence}, \emph{avg.contrib}, \emph{cor.contrib}, \emph{cor.cor}, and 
\emph{avg.cor}.
}
\description{
Calculates the modules preservation statistics using the pre-calculated 
module properties along with the correlation structure matrices.
}
\section{Design decision}{

 Most of the module presevation statistics are calculated as the correlation 
 of some network property between the \emph{discovery} and \emph{test} 
 datasets. The permutation procedure randomly samples from the \emph{test} 
 dataset, so it therefore makes sense to calculate the network properties in 
 the \emph{discovery} dataset once only. The code is thus split into 
 calculations of the network properties, and then subsequent calculation of 
 the statistics. The exception to this design are the correlation structure 
 based statistics as they require all entries of the correlation structure 
 matrices for that module. As modules can be quite large (e.g., 10,000 
 variables), this would be dramatically increase memory overhead to apply 
 this strategy.
}
\references{
\enumerate{
    \item{
      Langfelder, P., Luo, R., Oldham, M. C. & Horvath, S. \emph{Is my
      network module preserved and reproducible?} PLoS Comput. Biol. 
      \strong{7}, e1001057 (2011). 
    }
 }
}

