% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculatePreservation}
\alias{calculatePreservation}
\title{Calculates the preservation of a module between two networks using a wide
range of statistics.}
\usage{
calculatePreservation(refExpr, refAdj, testExpr, testAdj, moduleNodes)
}
\arguments{
\item{refExpr}{Expression matrix for the reference network.}

\item{refAdj}{Adjacency matrix for the reference network.}

\item{testExpr}{Expression matrix for the test network.}

\item{testAdj}{Adjacency matrix for the test network.}

\item{moduleNodes}{Nodes (genes) in the reference network for the module
whose preservation is being measured.}
}
\value{
A vector of preservation scores for the module.
}
\description{
Calculates the preservation of a module between two networks using a wide
range of statistics.
}
\details{
This is the actual guts of the calculation. This calculates the
preservation of each module based on the statistics outlined by Langfelder
et al. (See Reference below) and implemented in their R package WGCNA.
For each module the following are calculated depending on whether the
expression data, adjacency data, or both are provided:
\itemize{
 \item{Expression Data}{
   \itemize{
     \item{TODO:}{None Implemented yet}
   }
 }
 \item{Network Adjacencies}{
   \itemize{
     \item{\code{\link{meanAdj}}:}{
       The average edge weight of the module in the test network.
     }
   }
 }
}
}
\references{
Langfelder, Peter; Luo, Rui; Oldham, Michael C.; and Horvath,
   Steve. Is My Network Module Preserved and Reproducible?. PLoS
   Computational Biology, 2011.
}

