% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigMatrix-functions.R
\name{checkFinite}
\alias{checkFinite}
\title{Check if all entries of a `bigMatrix` are Finite}
\usage{
checkFinite(x)
}
\arguments{
\item{x}{a \code{\link{bigMatrix}}}
}
\description{
If there are non-finite entires (\code{NA}, \code{NaN}, \code{-Inf}, 
\code{Inf}), throw an exception.
}

