% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-user-input.R
\name{checkIsMatrix}
\alias{checkIsMatrix}
\title{Check whether an object is a 'matrix' or a 'disk.matrix'}
\usage{
checkIsMatrix(object)
}
\arguments{
\item{object}{object to check.}
}
\value{
throws an error or returns silently
}
\description{
Check whether an object is a 'matrix' or a 'disk.matrix'
}

