% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-user-input.R
\name{checkPlotArgs}
\alias{checkPlotArgs}
\title{Validate plot function arguments}
\usage{
checkPlotArgs(plotNodeNames, plotSampleNames, plotModuleNames, main,
  drawBorders, border.width, gaxt.line, saxt.line, maxt.line, legend.tick.size,
  laxt.line)
}
\arguments{
\item{plotSampleNames}{user input for the corresponding argument in the plot functions.}

\item{plotModuleNames}{user input for the corresponding argument in the plot functions.}

\item{main}{user input for the corresponding argument in the plot functions.}

\item{drawBorders}{user input for the corresponding argument in the plot functions.}

\item{border.width}{user input for the corresponding argument in the plot functions.}

\item{gaxt.line}{user input for the corresponding argument in the plot functions.}

\item{saxt.line}{user input for the corresponding argument in the plot functions.}

\item{maxt.line}{user input for the corresponding argument in the plot functions.}

\item{legend.tick.size}{user input for the corresponding argument in the plot functions.}

\item{laxt.line}{user input for the corresponding argument in the plot functions.}

\item{plotNodesNames}{user input for the corresponding argument in the plot functions.}
}
\description{
Simple typechecking for the extensive plot arguments
}

