% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkSetConsistency}
\alias{checkSetConsistency}
\title{Check consistency of all dataset lists}
\usage{
checkSetConsistency(datSets, netSets)
}
\arguments{
\item{datSets}{a list of data matrices (or \code{NULL})}

\item{netSets}{a list of networks (or \code{NULL})}
}
\description{
Throws an error if the dataset lists are not consistent.
}

