% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkSets}
\alias{checkSets}
\title{Check input data for consistency}
\usage{
checkSets(geneExpression, coexpression, adjacency, moduleAssignments)
}
\arguments{
\item{geneExpression}{a \code{\link{bigMatrix}} object, or a list of
\code{bigMatrix} objects, see details.}

\item{coexpression}{a \code{\link{bigMatrix}} object, or a list of
\code{bigMatrix} objects, see details.}

\item{adjacency}{a \code{\link{bigMatrix}} object, or a list of
\code{bigMatrix} objects, see details.}

\item{moduleAssignments}{a vector assigning genes to modules, or a list of
such vectors. See details.}
}
\description{
Make sure all objects are the right class, have the same dimensions, and
are ordered the same way across datasets.
}
\details{
#'
}

