% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CheckFinite}
\alias{CheckFinite}
\title{Check the elements of a `big.matrix`}
\usage{
CheckFinite(pDat)
}
\arguments{
\item{pDat}{SEXP container for the pointer to the 
\code{\link[bigmemory]{big.matrix}} to be checked.}
}
\description{
Are all the values finite?
}

