% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanTempDir}
\alias{cleanTempDir}
\title{Remove temporarily created bigMatrix objects}
\usage{
cleanTempDir()
}
\description{
Prevents the temporary directory from growing excessively large if using many
package functions in the same R session
}

