% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanupCluster}
\alias{cleanupCluster}
\title{De-register a parallel backend}
\usage{
cleanupCluster(cluster, predef)
}
\arguments{
\item{cluster}{registered cluster on a Windows machine}

\item{predef}{logical; was a pre-existing parallel backend used?}
}
\description{
De-register a parallel backend
}

