% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coexp_params}
\alias{coexp_params}
\title{Template parameters}
\arguments{
\item{discCoexp}{a \code{\link{bigMatrix}} containing the pairwise
coexpression for the \emph{discovery} dataset.}

\item{testCoexp}{a \code{\link{bigMatrix}} containing the pairwise
coexpression for the \emph{test} dataset.}

\item{discIndices}{indices corresponding to the network module in the matrix
of pairwise-coexpression in the \emph{discovery} dataset.}

\item{testIndices}{indices corresponding to the network module, or a random
  sample of the same size, in the matrix of pairwise-coexpression in the
  \emph{test} dataset.}
}
\description{
Template parameters to be imported into other function documentation. This
is not intended to be a stand-alone help file.
}

