% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coexpression.palette}
\alias{coexpression.palette}
\title{Color palette for coexpression and adjacency heatmaps}
\usage{
coexpression.palette()
}
\description{
RColorBrewer palette "RdYlBu" with the middle color replaced with white.
This gives a nicer contrast than the "RdBu" palette
}

