% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CoexpStats}
\alias{CoexpStats}
\title{Calculate the Correlation of Coexpression and Mean Sign-Aware Coexpression}
\usage{
CoexpStats(pCoexpD, discIndices, pCoexpT, testIndices)
}
\arguments{
\item{pCoexpD, pCoexpT}{SEXP containers for the pointers to the coexpression 
matrices for the \emph{discovery} and \emph{test} networks respectively.}

\item{discIndices, testIndices}{indices of the network subset in
the \emph{discovery} and \emph{test} networks respectively.}
}
\value{
A vector containing:
  \enumerate{
    \item{\emph{cor.coexp}:}{
      The correlation between the subset coexpression for both networks.
    }
    \item{\emph{mean.coexp}:}{
      The mean correlation density of the network subset.
    }
  }
}
\description{
Both of the coexpression statistics are calculated using all pairwise 
coexpression values in both the \emph{discovery} and \emph{test} datasets
respectively. For the other statistics, it makes sense to calculate the 
properties for the discovery network in advance to reduce calculation time
and memory. For these statistics this strategy doesn't make sense.
}
\references{
\enumerate{
    \item{
      Langfelder, P., Luo, R., Oldham, M. C. & Horvath, S. \emph{Is my 
      network module preserved and reproducible?} PLoS Comput. Biol. 
      \strong{7}, e1001057 (2011). 
    }
 }
}

