% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi-machine.R
\name{combineAnalyses}
\alias{combineAnalyses}
\title{Multi machine analysis}
\usage{
combineAnalyses(pres1, pres2)
}
\arguments{
\item{pres1, pres2}{lists returned by \code{\link{modulePreservation}}.}
}
\value{
A nested list containing the same elements as 
 \code{\link{modulePreservation}}.
}
\description{
This function takes the output from multiple runs of 
\code{\link{modulePreservation}}, combines their results, and returns a new
set of permutation test P-values. This is useful for parallelising 
calculations across multiple machines.
}
\details{
The input arguments must be for a single dataset comparison, i.e. the list
 must contain the elements \code{nulls} for the null distributions. If 
 multiple \code{discovery} or \code{test} datasets have been specified, you
 will need to extract the relevant list elements
}
\examples{
\dontrun{
data("NetRep")

# Set up input lists for each input matrix type across datasets. The list
# elements can have any names, so long as they are consistent between the
# inputs.
data_list <- list(discovery=discovery_data, test=test_data)
correlation_list <- list(discovery=discovery_correlation, test=test_correlation)
network_list <- list(discovery=discovery_network, test=test_network)
labels_list <- list(discovery=module_labels)

# Assess module preservation on two different machines to spread computational
# burden.
pres1 <- modulePreservation(
 data=data_list, correlation=correlation_list, network=network_list,
 moduleAssignments=labels_list, nPerm=500, discovery="discovery", 
 test="test"
)

pres2 <- modulePreservation(
 data=data_list, correlation=correlation_list, network=network_list,
 moduleAssignments=labels_list, nPerm=500, discovery="discovery", 
 test="test"
)

combined <- combineAnalyses(pres1, pres2)

}

}

