% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi-machine.R
\name{combineAnalysesInternal}
\alias{combineAnalysesInternal}
\title{Combine results of multiple permutation procedures}
\usage{
combineAnalysesInternal(pres1, pres2)
}
\arguments{
\item{pres1}{input from 'combineAnalyses'}

\item{pres2}{input from 'combineAnalyses'}
}
\value{
A combined module preservation analysis.
}
\description{
This function takes the module preservation analyses for two dataset 
comparisons and combines them.
}

