% Generated by roxygen2 (4.0.2): do not edit by hand
\name{combineNulls}
\alias{combineNulls}
\title{Combine null distribution calculations}
\usage{
combineNulls(...)
}
\arguments{
\item{...}{any number of three dimensional arrays holding the null
  distributions calculated by \code{\link{modulePreservation}}.}
}
\value{
A three dimensional array combined from the input. The rows correspond to
 the modules, columns to the module preservation statistics.
}
\description{
Combine null distribution calculations
}

