% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-defs.R
\name{common_params2}
\alias{common_params2}
\title{Template parameters}
\arguments{
\item{modules}{a vector of modules to apply the function to (see details).}

\item{discovery}{name or index denoting which dataset the module of interest
was discovered in (see details).}

\item{test}{name or index denoting which dataset to apply the function to 
(see details).}
}
\description{
Template parameters to be imported into other function documentation. This 
is not intended to be a stand-alone help file.
}

