% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ComparativeEdgeMatrix}
\alias{ComparativeEdgeMatrix}
\title{Combine edge weight matrices for a network derived from two different data sources.}
\usage{
ComparativeEdgeMatrix(edge.matrix1, edge.matrix2)
}
\arguments{
\item{edge.matrix1}{}

\item{edge.matrix2}{}
}
\value{
A edge.matrix matrix where the upper triangle corresponds to
  \code{edge.matrix1}, and the lower triangle corresponds to i
  \code{edge.matrix2}.
}
\description{
Primarily for use with \code{\link{PlotNetworkHeatmap}}.
}

