% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cor}
\alias{cor}
\title{Dispatch function for correlation methods}
\usage{
cor(x, y, method = "pearson")
}
\arguments{
\item{x}{a numeric vector or a matrix. If \code{y} is null, \code{x} must be
a matrix.}

\item{y}{a numeric vector or a matrix. If not given, correlations of columns
of \code{x} will be calculated.}

\item{method}{character string specifying the method to be used. If "pearson", the
 fast correlation function provided by the WGCNA package will be used. If
 "bicor", then the \code{\link[WGCNA]{bicor}} function will be used.
 Otherwise the standard \code{\link[stats]{cor}} will be used.}
}
\description{
Dispatch function for correlation methods
}

