% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-defs.R
\name{cor_params}
\alias{cor_params}
\title{Template parameters}
\arguments{
\item{discCor}{a \code{\link[=bigMatrix-class]{bigMatrix}} containing the
correlation structure among variables in the \emph{discovery} dataset.}

\item{testCor}{a \code{\link[=bigMatrix-class]{bigMatrix}} containing the
correlation structure among variables in the \emph{test} dataset.}

\item{discIndices}{indices corresponding to the network module in 
\code{discCor}.}

\item{testIndices}{indices corresponding to the network module in
\code{testCor}.}
}
\description{
Template parameters to be imported into other function documentation. This 
is not intended to be a stand-alone help file.
}

