% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{correlation.palette}
\alias{correlation.palette}
\title{Color palette for correlation heatmaps}
\usage{
correlation.palette()
}
\description{
RColorBrewer palette "RdYlBu" with the middle color replaced with white.
This gives a nicer contrast than the "RdBu" palette
}

