% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CorStats}
\alias{CorStats}
\title{Calculate the correlation based statistics}
\usage{
CorStats(pCorD, discIndices, pCorT, testIndices)
}
\arguments{
\item{pCorD, pCorT}{SEXP containers for the pointers to the correlation 
structure matrices for the \emph{discovery} and \emph{test} networks.}

\item{discIndices, testIndices}{indices of the network subset in
the \emph{discovery} and \emph{test} networks respectively.}
}
\value{
A vector containing:
  \enumerate{
    \item{\emph{cor.discovery}:}{
      A flattened vector of the module's correlation structure in the 
      \emph{discovery} dataset.
    }
    \item{\emph{cor.test}:}{
      A flattened vector of the module's correlation structure in the 
      \emph{test} dataset.
    }
    \item{\emph{mean.cor}:}{
      The mean sign-aware correlation density of the network module.
    }
  }
}
\description{
Both of the correlation statistics are calculated using all pairwise 
correlation values in both the \emph{discovery} and \emph{test} datasets.
For the other statistics, it makes sense to calculate the 
properties for the discovery network in advance to reduce calculation time
and memory. However, for the correaltion statistics this strategy doesn't 
make sense since we'd have to store a huge component of the discovery 
correlation structure.
}
\references{
\enumerate{
    \item{
      Langfelder, P., Luo, R., Oldham, M. C. & Horvath, S. \emph{Is my 
      network module preserved and reproducible?} PLoS Comput. Biol. 
      \strong{7}, e1001057 (2011). 
    }
 }
}

