% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-defs.R
\name{dat_param}
\alias{dat_param}
\title{Template parameters}
\arguments{
\item{sge}{a \code{\link[=bigMatrix-class]{bigMatrix}} containing scaled 
numeric data. Assumes columns correspond to variables (e.g. genes, 
microbial operational taxonomic unit) and rows correspond to samples.}
}
\description{
Template parameters to be imported into other function documentation. This 
is not intended to be a stand-alone help file.
}

