% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{data.palette}
\alias{data.palette}
\title{Color palette for data heatmaps and module summary vector plots}
\usage{
data.palette()
}
\description{
RColorBrewer palette "PRGn" with a white middle color.
}

