% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DataProps}
\alias{DataProps}
\title{Get the module summary vector and the proportion of variance it explains}
\usage{
DataProps(pDat, subsetIndices)
}
\arguments{
\item{pDat}{SEXP container for the pointer to a scaled version of the 
data matrix used to construct the network.}

\item{subsetIndices}{indices of the network subset of interest in 
\code{pDat}.}
}
\value{
A list containing:
 \enumerate{
  \item{\emph{"moduleSummary"}:}{
    The module's summary profile (see details).
  }
  \item{\emph{"nodeContribution"}:}{
    The contribution of each node to the module's summary profile (see 
    details).
  }
  \item{\emph{"moduleCoherence"}:}{
    The proportion of the variance in the data explained by the module's 
    summary profile (see details).
  }
 }
}
\description{
Get the module summary vector and the proportion of variance it explains
}
\details{
First, the module summary profile ('moduleSummary') is calculated as the
 first eigenvector of a principal component analysis of the variables
 composing the module of interest. The orientation of the eigenvector is
 modified so that its sign is in the same direction as the module on
 average. I.e. for gene expression data this is the "module eigengene"
 \emph{(1)}.
 
 Each node's contribution to the summary profile ('nodeContribution') is
 quantified as the correlation between each variable composing the module
 and the module's summary profile. For weighted gene coexpression networks,
 this is typically referred to as the 'module membership' \emph{(1)}.
 
 The The proportion of module variance explained by the summary profile
 ('moduleCoherence') is quantified as the average square of the
 'nodeContribution' \emph{(1)}.
}
\references{
\enumerate{
    \item{
      Langfelder, P., Luo, R., Oldham, M. C. & Horvath, S. \emph{Is my
      network module preserved and reproducible?} PLoS Comput. Biol. 
      \strong{7}, e1001057 (2011). 
    }
 }
}

