% Generated by roxygen2 (4.0.1): do not edit by hand
\name{DataSummary}
\alias{DataSummary}
\title{Network subset eigenvector and proportion of variance explained in C++}
\usage{
DataSummary(pDat, pScaledDat, subsetIndices)
}
\arguments{
\item{pDat}{SEXP container for the pointer to the data matrix used in
network construction.}

\item{pScaledDat}{SEXP container for the pointer to a scaled version of the
data matrix used to construct the network.}

\item{subsetIndices}{indices of the network subset of interest in
  \code{pDat}.}
}
\value{
A list whose first element is the subset membership for each node (see
 details), and whose second element is the proportion of the variance
 explained by the subset's summary vector (see details).
}
\description{
Network subset eigenvector and proportion of variance explained in C++
}
\details{
First, a summary vector is calculated for the network subset from the
 underlying data. This is the first right singular vector from a singular
 value decomposition (also the eigenvector of the first principal component
 \emph{(1)}). The sign of the returned eigenvector is modified to match the
 average of \code{pDat}. This is to match the behaviour of
 \emph{moduleEigengenes} in the \code{WGCNA} package.

 Using this summary vector, the subset membership of each node is quantified
 as the correlation between that node's data, and the summary vector.

 The proportion of variance explained by this summary vector is quantified
 as the average square of the subset memberships for all nodes in the
 network subset.

 The two returned properties are bundled together into one function because
 the calculation of the proportion of variance requires much of the same
 underlying intermediate calculations that obtaining the first eigenvector
 requires.
}
\references{
\enumerate{
    \item{
      Langfelder, P., Luo, R., Oldham, M. C. & Horvath, S. \emph{Is my
      network module preserved and reproducible?} PLoS Comput. Biol.
      \strong{7}, e1001057 (2011).
    }
 }
}

