% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{expression.palette}
\alias{expression.palette}
\title{Color palette for expression and summary expression plots}
\usage{
expression.palette()
}
\description{
RColorBrewer palette "PRGn" with a white middle color.
}

