% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{fastModPres}
\alias{fastModPres}
\alias{fastModPres-package}
\title{Fast permutation procedure for testing network module replication}
\description{
Functions for assessing the replication/preservation of network
topology for weighted gene coexpression network modules in one or more
independent datasets through permutation testing.
}
\details{
The main function for this package is \code{\link{modulePreservation}}. Other
useful functions include \code{\link{networkProperties}} for calculating the
topological properties of a module, and \code{\link{plotModule}} and other
plotting functions for visualising a module.
}

