% Generated by roxygen2 (4.0.1): do not edit by hand
\name{foreach}
\alias{\%maybe_do_par\%}
\alias{foreach}
\title{Maybe do parallel}
\usage{
obj \%maybe_do_par\% ex
}
\arguments{
\item{obj}{\code{foreach} object used to control the evaluation of \code{ex}.}

\item{ex}{the R expression to evaluate.}
}
\description{
Suppresses warnings from \code{\link[foreach]{\%dopar\%}} when no parallel
backend is registered, using \code{\link[foreach]{\%do\%}} instead.
}
\seealso{
\code{\link{foreach}}
}

