% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-user-input.R
\name{formatDataList}
\alias{formatDataList}
\title{Format the optional 'data' argument list}
\usage{
formatDataList(data, nDatasets, datasetNames)
}
\arguments{
\item{data}{user provided input to 'data' argument of package functions.}

\item{nDatasets}{total number of datasets}

\item{datasetNames}{names of the datasets}
}
\value{
A formatted list of gene expression data
}
\description{
Allows the user to set \code{data} to \code{NULL} without the data
causing cascading input-check errors.
}

