% Generated by roxygen2 (4.0.2): do not edit by hand
\name{formatGeneExpression}
\alias{formatGeneExpression}
\title{Format the gene expression lists}
\usage{
formatGeneExpression(geneExpression, nDatasets, datasetNames)
}
\arguments{
\item{geneExpression}{geneExpression data}

\item{nDatasets}{total number of datasets}

\item{datasetNames}{names of the datasets}
}
\value{
A formatted list of gene expression data
}
\description{
Allows the user to set \code{geneExpression} to \code{NULL} without the data
causing cascading input-check errors.
}

