% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-user-input.R
\name{formatModuleAssignments}
\alias{formatModuleAssignments}
\title{Format moduleAssignments list}
\usage{
formatModuleAssignments(moduleAssignments, discovery, nDatasets, datasetNames,
  nDiscVars, discVarNames)
}
\arguments{
\item{moduleAssignments}{input provided by the user}

\item{discovery}{vector of discovery networks provided by the user}

\item{nDatasets}{number of datasets}

\item{datasetNames}{names of the datasets}

\item{nDiscVars}{number of variables in the discovery dataset. 
Only required if \code{moduleAssignments} is missing.}

\item{discVarNames}{varaibles in the discovery dataset. Only 
required if \code{moduleAssignments} is missing.}
}
\value{
List structure for moduleAssignments for internal calculations.
}
\description{
If module discovery has not been performed for all datasets, it may be
easier for the user to provide a simplified list structure. This function
unifies it to the same structure as the other datasets.
}

