% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ge_param}
\alias{ge_param}
\title{Template parameters}
\arguments{
\item{sge}{a \code{\link{bigMatrix}} containing scaled gene expression data.
  Assumes columns are genes/probes, rows are samples, and that probes have
  been scaled (This can be done using \code{\link{scaleBigMatrix}}).}
}
\description{
Template parameters to be imported into other function documentation. This
is not intended to be a stand-alone help file.
}

