% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetDiag}
\alias{GetDiag}
\title{big.matrix diagonals}
\usage{
GetDiag(pBigMat)
}
\arguments{
\item{pBigMat}{a \code{numeric} \code{\link[bigmemory]{big.matrix}}}
}
\value{
a vector containing the diagonal of \code{pBigMat}
}
\description{
big.matrix diagonals
}

