% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getModuleMidPoints}
\alias{getModuleMidPoints}
\title{Get module mid-points on the x-axis}
\usage{
getModuleMidPoints(mas)
}
\arguments{
\item{mas}{ordered subset of the moduleAssignments vector}
}
\value{
a vector of positions on the x-axis indicating the centre of a module
}
\description{
Get module mid-points on the x-axis
}

