% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getModuleVarsUnsorted}
\alias{getModuleVarsUnsorted}
\title{Get variables composing a module}
\usage{
getModuleVarsUnsorted(moduleAssignments, modules, discovery = 1)
}
\arguments{
\item{moduleAssignments}{a vector of the module assignment for each variable
 in the discovery dataset. If there are multiple discovery datasets 
then this argument should be a list of such vectors.}

\item{modules}{a vector of modules to apply the function to.}

\item{discovery}{name or index denoting which dataset the module of
interest was discovered in.}
}
\description{
Get the variables composing the user specified modules in the order they 
appear in the 'moduleAssignments' vector for the discovery dataset.
}

