% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getGenes}
\alias{getGenes}
\title{Get unordered genes}
\usage{
getGenes(moduleAssignments, modules, discovery = 1)
}
\arguments{
\item{moduleAssignments}{a vector assigning genes to modules, or a list of 
such vectors.}

\item{modules}{a vector of modules to apply the function to.}

\item{discovery}{name or index denoting which dataset the module of
interest was discovered in.}
}
\description{
Get the genes for the user specified modules in the order they appear in 
the moduleAssignments vector for the discovery dataset.
}
\details{
Genes are retrieved from the discovery dataset, rather than the test 
 dataset, because some genes from the discovery dataset may not appear in
 the test dataset.
}

