% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gradient.bar}
\alias{gradient.bar}
\title{Plot a Gradient along a Bar}
\usage{
gradient.bar(range = c(0, 1), break.points = NULL, nBins = NULL,
  col = NULL, col.gradient = NULL, border = par("fg"), lines = NA,
  direction = "x", main = "", bin.lab = NULL)
}
\arguments{
\item{range}{The range of the plot in the direction of the gradient (e.g.
xlim if the gradient is plotted over the x axis). Useful if the
\code{break.points} specified by the user have been determined by some
other function over some range.}

\item{break.points}{A vector of locations where each step of the gradient
transitions into the next.}

\item{nBins}{Splits the gradient bar into \code{nBins} evenly spaced bins.
Ignored if \code{break.points} are given.}

\item{col}{A vector of colors used to color each bin in the gradient bar.}

\item{col.gradient}{Colors to generate a gradient between. For example,
c("white", "red"), or, c("red", "white", "blue"). See
\code{help(colorRampPalette)} for more details. Ignored if \code{col} is
provided.}

\item{border}{The color of the border around the entire bar, or NA for no
border.}

\item{lines}{The color(s) of the lines to split each gradient step, or NA
to not draw lines.}

\item{direction}{The direction to render the gradient, must be one of 'x' or
'y'. To reverse the direction the gradient is draw on an axis, simply
reverse the color vectors provided to either \code{code} or
\code{col.gradient}.}

\item{main}{Text to render above the gradient bar.}

\item{bin.lab}{Text labels for each of the gradient bins. Will either plot
   the labels next to the bins, or the breaks between bins, depending on how
   many labels are provided. Labels are rendered either on the left, or
   bottom of the graph depending on the direction of the gradient. Useful for
   labelling legends}
}
\description{
\code{gradient.bar} plots a gradient along a bar in either the x or y
directions, using user specified break points for each step in the gradient.
This works by plotting the gradient over the entire dimensions of a plot
window.
}
\examples{
# Intervals of size 2^x
gradient.bar(range=c(0, 1024), break.points=2^(1:10),
  col.gradient=c("red", "white"))

# Plot a legend next to a bar plot
layout(matrix(c(1, 2), nrow=1), widths=c(0.8, 0.2))
barplot(1:10, col=1:10)
gradient.bar(nBins=10, col=1:10, direction='y', main="legend",
             bin.lab=letters[1:10], lines="black")

# Rainbow box
gradient.bar(nBins=256, col=rainbow(256))
}
\keyword{gradient}
\keyword{legend}

