% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-defs.R
\name{ind_param}
\alias{ind_param}
\title{Template parameters}
\arguments{
\item{moduleIndices}{indices for the network module of interest in the
supplied data matrix.}
}
\description{
Template parameters to be imported into other function documentation. This 
is not intended to be a stand-alone help file.
}

