% Generated by roxygen2 (4.0.2): do not edit by hand
\name{insert.nas}
\alias{insert.nas}
\title{Insert NAs into a vector at specified positions}
\usage{
insert.nas(vec, na.indices)
}
\arguments{
\item{vec}{vector to insert NAs to.}

\item{na.indices}{indices the NAs should be located at in the final vector.}
}
\value{
The vector with NAs inserted in the correct positions.
}
\description{
Useful for inserting NAs into the correct positions when examining module
probes that do not exist in the test dataset.
}

