% Generated by roxygen2 (4.0.1): do not edit by hand
\name{kIMR}
\alias{kIMR}
\title{Node Connectivity (weighted degree), R Implementation}
\usage{
kIMR(adjacency, subsetIndices, allNodes, undirected)
}
\arguments{
\item{adjacency}{Adjacency matrix representation of the network}

\item{subsetIndices}{row/column indices in the provided \code{adjacency} that
correspond to the network subset of interest.}

\item{allNodes}{Logical; if \code{FALSE}, the connectivity is calculated
based only on nodes in the module, otherwise a module's connectivity to all
nodes is calculated.}

\item{undirected}{Logical; if \code{TRUE}, only the lower half of the
symmetric adjacency matrix will be used in the calculations.}
}
\value{
The connectivity (weighted degree) for each node in the network
subset
}
\description{
Used to unit test the C++ functionality of \code{\link{kIM}}.
}

