% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadIntoRAM}
\alias{loadIntoRAM}
\title{Get the matrix in RAM}
\usage{
loadIntoRAM(x)
}
\arguments{
\item{x}{a \code{'matrix'} or \code{'big.matrix'}}
}
\value{
a \code{'matrix'}
}
\description{
If \code{x} is a \code{'big.matrix'} convert it to a matrix. This loads it 
into regular memory rather than shared memory, since shared memory is not
compatible with distributed file systems (i.e. multi-node clusters). If
\code{x} is already a \code{'matrix'}, just return as is.
}

