% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeModuleAssignments}
\alias{makeModuleAssignments}
\title{Format moduleAssignments list}
\usage{
makeModuleAssignments(moduleAssignments, discovery, nDatasets, datasetNames,
  nDiscGenes, discGeneNames)
}
\arguments{
\item{moduleAssignments}{input provided by the user}

\item{discovery}{vector of discovery networks provided by the user}

\item{nDatasets}{number of datasets}

\item{datasetNames}{names of the datasets}

\item{nDiscGenes}{number of genes in the discovery dataset. Only required if
moduleAssignments is missing.}

\item{discGeneNames}{names of the genes in the discovery dataset. Only
 required if moduleAssignments is missing.}
}
\value{
List structure for moduleAssignments for internal calculations.
}
\description{
If module discovery has not been performed for all datasets, it may be
easier for the user to provide a simplified list structure. This function
unifies it to the same structure as the other datasets.
}

