% Generated by roxygen2 (4.0.1): do not edit by hand
\name{matchsub}
\alias{\%nin\%}
\alias{\%sub_in\%}
\alias{matchsub}
\title{Value Matching and Subsetting}
\usage{
x \%nin\% table

x \%sub_in\% table
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched. \link[base]{Long vectors}
are supported.}

\item{table}{vector or \code{NULL}: the values to be matched against.
\link[base]{Long vectors} are not supported.}
}
\description{
This set of functions provides shortcuts for value matching and subsetting,
on top of the functionality provided by \code{\link[base]{\%in\%}}.

\code{\%nin\%} returns a logical vector indicating if elements of \code{x}
are not in \code{table}, This is the opposite of \code{\%in\%}.

\code{\%sub_in\%} returns the elements \code{x} that are \code{\%in\%}
\code{table} rather than a logical vector.

\code{\%sub_nin\%} returns the elements \code{x} that are \code{\%nin\%}
\code{table} rather than a logical vector.
}

