% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MaxThreads}
\alias{MaxThreads}
\title{Get the maximum number of concurrent threads supported by this machine}
\usage{
MaxThreads()
}
\description{
Get the maximum number of concurrent threads supported by this machine
}

