% Generated by roxygen2 (4.0.1): do not edit by hand
\name{meanAdjR}
\alias{meanAdjR}
\title{Mean Adjacency of a Network, R Implementation.}
\usage{
meanAdjR(adjacency, susbetIndices)
}
\arguments{
\item{adjacency}{Adjacency matrix.}

\item{subsetIndices}{indices of the subset of nodes to calculate the mean
adjacency on.}
}
\value{
a single numeric value.
}
\description{
Used to unit test the functionality of \code{\link{meanAdj}}.
}

