% Generated by roxygen2 (4.0.1): do not edit by hand
\name{modulePreservation}
\alias{modulePreservation}
\title{Calculates pairwise between reference and test networks}
\usage{
modulePreservation(expressionSets = NULL, adjacencySets = NULL,
  referenceSets, testSets, moduleLabels, ignoreCol = NULL,
  nPermutations = 200, verbose = TRUE, logFile = "log.txt", save = TRUE,
  permutedFile = "permutedStatistics.rda")
}
\arguments{
\item{expressionSets}{}

\item{adjacencySets}{}

\item{referenceSets}{}

\item{testSets}{}

\item{nPermutations}{}

\item{moduleLabels}{}

\item{ignoreCol}{A vector of \code{moduleLabels} colors to ignore when
calculating preservation.}

\item{verbose}{Logical; If \code{TRUE}, outputs message to file.}

\item{logFile}{File to send output to, if \code{verbose} is \code{TRUE}.}

\item{savePermuted}{}

\item{permutedFile}{}
}
\value{
a
}
\description{
Description goes here
}

