% Generated by roxygen2 (4.0.1): do not edit by hand
\name{monitorProgress}
\alias{monitorProgress}
\title{Monitor Parallel Progress}
\usage{
monitorProgress(updateFreq = 2)
}
\arguments{
\item{updateFreq}{number or seconds to wait between chunk switches.}
}
\description{
Monitor the progress of null distribution calculation.
}
\details{
This must be run in a new R session.
}
\note{
If you see the message \emph{"Waiting for parallel code to start..."} and
 \code{\link{netRep}} is already calculating the null distributions, your R
 session has been initialized in the wrong directory.
}

