% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-defs.R
\name{net_param}
\alias{net_param}
\title{Template parameters}
\arguments{
\item{adj}{a \code{\link[=bigMatrix-class]{bigMatrix}} containing an
adjacency matrix representation of the network (i.e. an \eqn{n * n} matrix 
containing the edge weights between each pair of variables).}
}
\description{
Template parameters to be imported into other function documentation. This 
is not intended to be a stand-alone help file.
}

