% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream-analysis.R
\name{netPropsInternal}
\alias{netPropsInternal}
\title{Internal function for calculating 'networkProperties'}
\usage{
netPropsInternal(scaledData, correlation, network, moduleAssignments, modules,
  discovery, test, nDatasets, datasetNames, verbose)
}
\arguments{
\item{scaledData}{scaled data generated by \code{'processInput'}.}

\item{correlation}{\code{'correlation'} after processing by
\code{'processInput'}.}

\item{network}{\code{'network'} after processing by \code{'processInput'}.}

\item{moduleAssignments}{\code{'moduleAssignments'} after processing by
\code{'processInput'}.}

\item{modules}{\code{'modules'} after processing by \code{'processInput'}.}

\item{discovery}{\code{'discovery'} after processing by
\code{'processInput'}.}

\item{test}{\code{'test'} after processing by \code{'processInput'}.}

\item{nDatasets}{a vector containing the total number of input datasets, 
returned by \code{'processInput'}.}

\item{datasetNames}{a vector of dataset names returned by
\code{'processInput'}.}

\item{verbose}{logical; should progress be reported? Default is \code{TRUE}.}
}
\description{
This function is used by several package functions. It assumes that all user
input has been processed already by \code{\link{processInput}}. This allows
for function-specific checking (i.e. failing early where the \code{'data'} is
required), while avoiding duplication of time-intensive checks 
(e.g. \code{\link{checkFinite}}) and data duplication (e.g. through
\code{\link{scaleBigMatrix}}).
}

