% Generated by roxygen2 (4.0.1): do not edit by hand
\name{networkReplication}
\alias{networkReplication}
\title{Are my networks preserved and reproducible?}
\usage{
networkReplication(datSets, netSets, discoverySets, testSets, nodeLabelSets,
  ignoreSets, nPerm = 10000, verbose = TRUE, simplify = TRUE)
}
\arguments{
\item{datSets}{A list of \code{\link{big.matrix}} the underlying data the
networks are calculated from.}

\item{netSets}{A list of \code{\link{big.matrix}}}

\item{discoverySets}{A numeric vector}

\item{testSets}{A numeric vector}

\item{nodeLabelSets}{A list of vectors, one for each discovery data set,
assigning each node in the discovery dataset to a subnetwork.}

\item{ignoreSets}{An optional list of vectors, one for each discovery data
set, specifying which subnetworks to ignore.}

\item{nPerm}{number of permutations to use.}

\item{verbose}{logical indicating whether to print detailed output about the
function's execution.}

\item{simplify}{logical. If \code{FALSE} the structure of the returned list
will be regular: that is, a list of lists, where the top level indicates a
summary network.}
}
\description{
Are my networks preserved and reproducible?
}

