% Generated by roxygen2 (4.0.1): do not edit by hand
\name{networkReplicationCore}
\alias{networkReplicationCore}
\title{Core Function for Assessing Network Replication}
\usage{
networkReplicationCore(datSets = NULL, netSets = NULL, nodeLabelSets,
  discovery, test, buildNetFun, ignoreSets, includeSets)
}
\arguments{
\item{datSets}{\code{NULL}, or a list of \code{\link{big.matrix}}: one for
each dataset.}

\item{netSets}{\code{NULL}, or a list of \code{\link{big.matrix}}: one for
each dataset. Alternatively, if the list elements are \code{NULL}, then the
networks will be dynamically calculated using the provided
\code{buildNetFun} on the corresponding \code{datSet}.}

\item{nodeLabelSets}{a list, whose elements are \code{NULL} for each
\emph{test} dataset, and a vector for each \emph{discovery} dataset
assigning each node to a sub-network/cluster/module/component.}

\item{discovery}{a numeric vector indicating which elements of
\code{datSets} and/or \code{netSets} are to be treated as the
\emph{discovery} datasets.}

\item{test}{a numeric vector indicating which elements of \code{datSets}
and/or \code{netSets} are to be treated as the \emph{test} datasets.}

\item{buildNetFun}{A function for constructing a network from the
\code{datSets}. This is used only where a pre-constructed network is not
provided in \code{netSets}.}

\item{ignoreSets}{An optional list, where the elments for each
\emph{discovery} data set are vectors specifying which sub-networks to
skip.}

\item{includeSets}{An optional list, where the elments for each
  \emph{discovery} data set are vectors specifying which sub-networks to
  include.}
}
\description{
This function provides the main functionality, but requires well formed
input. A wrapper function has been provided for usability purposes, see
\code{\link{networkReplication}} for details.
}
\details{
Any function argument ending in "Sets" expects a list, where each element
 corresponds to a dataset. The order of the datasets should match across all
 arguments.

 If \code{datSets} is \code{NULL}, then this function assumes a full list of
 \code{netSets} has been provided for each dataset of interest.

 \code{ignoreSets} and \code{includeSets} should not overlap.
}

