% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream-analysis.R
\name{nodeOrderInternal}
\alias{nodeOrderInternal}
\title{Internal use function for calculating node order}
\usage{
nodeOrderInternal(props, orderModules, simplify, verbose, na.rm)
}
\arguments{
\item{props}{network properties to calculate order from.}

\item{orderModules}{logical; are we ordering the modules by similarity?}

\item{simplify}{logical; are we simplifying the output?}

\item{verbose}{logical; is verbose printing turned on?}

\item{na.rm}{logical; remove missing nodes?}
}
\value{
list structure of ordered nodes.
}
\description{
Used by plotting functions: assumes user input has been sanitised already
}

