% Generated by roxygen2 (4.0.2): do not edit by hand
\name{orderModule}
\alias{orderModule}
\title{Order genes and modules in a network}
\usage{
orderModule(adjacency, module.labels = NULL, summary.exp = NULL)
}
\arguments{
\item{adjacency}{Matrix of gene adjacencies}

\item{module.labels}{Optional vector of module assignments for each gene.}

\item{summary.exp}{Optional \code{data.frame} of summary expression profiles
 for each network subset.}
}
\value{
a vector of ordered nodes.
}
\description{
Clusters network modules by summary expression profile similarity, and
order genes within each subset by connectivity.
}

