% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-defs.R
\name{orderModules_param}
\alias{orderModules_param}
\title{Template parameters}
\arguments{
\item{orderModules}{logical; if \code{TRUE} modules ordered by clustering 
their summary vectors. If \code{FALSE} modules are returned in the order
provided.}
}
\description{
Template parameters to be imported into other function documentation. This 
is not intended to be a stand-alone help file.
}
\keyword{internal}
