% Generated by roxygen2 (4.0.2): do not edit by hand
\name{orderSamples}
\alias{orderSamples}
\title{Order gene expression samples based on the summary expression profile}
\usage{
orderSamples(gene.expr)
}
\arguments{
\item{gene.expr}{matrix of gene expression data. Expects columns to be the
genes.}
}
\value{
a vector of ordered nodes.
}
\description{
Order gene expression samples based on the summary expression profile
}

