% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-defs.R
\name{par_param}
\alias{par_param}
\title{Template parameters}
\arguments{
\item{nCores}{number of cores to parallelise the calculation of network 
properties over. Ignored if the user has already registered a parallel 
backend.If \code{NULL} (default) the maximum number of cores on the machine
will be used.}
}
\description{
Template parameters to be imported into other function documentation. This 
is not intended to be a stand-alone help file.
}

