% Generated by roxygen2 (4.0.2): do not edit by hand
\name{permutation}
\alias{perm.test}
\alias{permutation}
\alias{permuted}
\alias{requiredPower}
\title{Distribution from Permutations}
\usage{
perm.test(permuted, observed, subsetSize, totalSize, order = TRUE,
  alternative = "greater")

requiredPower(alpha)
}
\arguments{
\item{permuted}{vector of values making up the empirical distribution.}

\item{observed}{the observed value of the test statistic.}

\item{subsetSize}{the size of the network subset the null distribution is
drawn for.}

\item{totalSize}{the size of the whole network}

\item{order}{logical; does the order of nodes in the permutation affect the
value of the test statistic?}

\item{alternative}{a character string specifying the alternative hypothesis,
 must be one of "greater" (default), "less", or "two.sider".
 You can specify just the initial letter.}

\item{alpha}{desired significance threshold.}
}
\value{
The minimum number of permutations required to detect any significant
 associations at the provided \code{alpha}. The minimum p-value will always
 be smaller than \code{alpha}.
}
\description{
Distribution function, quantile function, and random generation
 for an empirical distribution provided by the user in the \code{permuted}
 argument. These distributions are typically drawn through permutation
 testing of some statistic.

\code{requiredPower}:  how many permutations do I need to be able to detect
 significance at a given threshold \code{alpha}?
}
\details{
P-values are calculated by \code{pperm} using proportions on the provided
 distribution (\emph{permuted}). An observation, \code{q}'s, p-value
 is the proportion of \emph{permuted} whom are more extreme than \code{q}.
 This is accurate strictly when:
 \enumerate{
   \item Each permutation is independent.
   \item Permutations are sampled without replacement.
 }
 While these assumptions may not necessarily always be true, for our purposes
 they are a reasonable approximation. If \code{q} is more extreme than the
 \code{permuted}, the tail approximation method from \emph{(2)} can be used
 to estimate the p-value of \code{q} by setting \code{tail.approx} to
 \code{TRUE}.

 It is rarely possible to give an accurate value for \code{qperm} since there
 is no theoretical distribution to draw from. Instead the closest possible
 observation from the data is returned.
}
\note{
It is possible for the quantile \code{p} to fall exactly halfway
 between two observations from the \emph{permuted} distribution. In this case
 both data points are returned and a warning is generated. It is up to the
 user to choose which observation to take: the conservative approach is to
 choose the first observation if \code{p} < 0.5, or the second observation if
 \code{p} > 0.5.
}
\references{
\enumerate{
    \item{
      Phipson, B. & Smyth, G. K. \emph{Permutation P-values should never be
      zero: calculating exact P-values when permutations are randomly drawn.}
      Stat. Appl. Genet. Mol. Biol. \strong{9}, Article39 (2010).
    }
  }
}

