% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PermutationProcedure}
\alias{PermutationProcedure}
\title{Multithreaded permutation procedure for module preservation statistics}
\usage{
PermutationProcedure(dData, dCorr, dNet, tData, tCorr, tNet, moduleAssignments,
  modules, nPermutations, nCores)
}
\arguments{
\item{dData}{data matrix from the \emph{discovery} dataset.}

\item{dCorr}{matrix of correlation coefficients between all pairs of 
variables/nodes in the \emph{discovery} dataset.}

\item{dNet}{adjacency matrix of network edge weights between all pairs of 
nodes in the \emph{discovery} dataset.}

\item{tData}{data matrix from the \emph{test} dataset.}

\item{tCorr}{matrix of correlation coefficients between all pairs of 
variables/nodes in the \emph{test} dataset.}

\item{tNet}{adjacency matrix of network edge weights between all pairs of 
nodes in the \emph{test} dataset.}

\item{moduleAssignments}{a named character vector containing the module 
each node belongs to in the discovery dataset.}

\item{modules}{a character vector of modules for which to calculate the 
module preservation statistics.}

\item{nPermutations}{the number of permutations from which to generate the
null distributions for each statistic.}

\item{nCores}{the number of cores that the permutation procedure may use.}
}
\value{
a list containing a matrix of observed test statistics, and an
  array of null distribution observations.
}
\description{
Multithreaded permutation procedure for module preservation statistics
}
\details{
\subsection{Input expectations:}{
  Note that this function expects all inputs to be sensible, as checked by
  the R function 'checkUserInput' and processed by 'modulePreservation'. 
  
  These requirements are:
  \itemize{
  \item{The ordering of node names across 'dData', 'dCorr', and 'dNet' is
        consistent.}
  \item{The ordering of node names across 'tData', 'tCorr', and 'tNet' is
        consistent.}
  \item{The columns of 'dData' and 'tData' are the nodes.}
  \item{'dCorr', 'dNet', 'tCorr', and 'tNet' are square matrices, and their
        rownames are identical to their column names.}
  \item{'moduleAssigments' is a named character vector, where the names
        represent node labels found in the test dataset (i.e. 'tNet').
        The vector must not include nodes that are not present in the 
        test dataset.}
  \item{The labels in 'modules' must all be present in 'moduleAssignments'.}
  \item{'nPermutations' is a single number, greater than 0.}
  \item{'nCores' is a single number, greater than 0. Note, this number must
        not be larger than the number of cores on your machine, or the 
        number of cores allocated to your job!}
  }
}
}

