% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkgReqCheck}
\alias{pkgReqCheck}
\title{Silently check and load a package into the namespace}
\usage{
pkgReqCheck(pkg)
}
\arguments{
\item{pkg}{name of the package to check}
}
\value{
logical; \code{TRUE} if the package is installed and can be loaded.
}
\description{
Silently check and load a package into the namespace
}
\keyword{internal}

