% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotBar}
\alias{plotBar}
\title{Custom bar plot function}
\usage{
plotBar(heights, heights.lim, mas, cols, bar.width = 1, drawBorders = FALSE,
  na.col = "#bdbdbd", xaxt = TRUE, plotModuleNames = TRUE, main = "",
  xaxt.line = -0.5, maxt.line = 3, ylab = "", border.width = 2)
}
\arguments{
\item{heights}{heights of the bars.}

\item{heights.lim}{limits of the height axis.}

\item{mas}{ordered subset of the moduleAssignments vector}

\item{cols}{colors of each bar.}

\item{bar.width}{value between 0 and 1 controlling the proportion of space
taken by each bar.}

\item{drawBorders}{logical; if \code{TRUE} a border is drawn around each bar.}

\item{na.col}{color of missing values to plot.}

\item{xaxt}{logical; If \code{TRUE}, the names of \code{heights} will be
rendered underneath the bar chart}

\item{plotModuleNames}{logical; if \code{TRUE} the names of the modules are
plotted along the x axis.}

\item{main}{title for the plot.}

\item{xaxt.line}{the number of lines into the margin at which the x axis
labels will be drawn.}

\item{maxt.line}{the number of lines into the margin at which the module
names will be drawn.}

\item{ylab}{label for the yaxis}

\item{border.width}{line width for borders.}
}
\description{
Plot bars around 0
}

