% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotCoexpLegend}
\alias{plotCoexpLegend}
\title{Plot legend for coexpression}
\usage{
plotCoexpLegend(heatmap.gradient, nBins = 11,
  title = "Pearson\\nCorrelation", cex.title = 1.4, cex.legend = 1)
}
\arguments{
\item{heatmap.gradient}{A vector of colors that was used in
\code{\link{plotCoexpression}}}

\item{nBins}{number of bins to show on the legend.}
}
\description{
Plot legend for coexpression
}

