% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotGeneNames}
\alias{plotGeneNames}
\title{Render gene names for coexpression plot}
\usage{
plotGeneNames(gene.labels, bot.mar = 5, cex.axis = 1)
}
\arguments{
\item{gene.labels}{labels for each gene to render}

\item{bot.mar}{bottom margin}

\item{cex.axis}{cex for the axis text}
}
\description{
Create panel containing the axis with just the gene names. Adjust bot.mar to
get output correct.
}

