% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotKIM}
\alias{plotKIM}
\title{Plot the Intramodular Connectivity}
\usage{
plotKIM(adjacency, module.labels, col = "orange", cex.axis = 1,
  cex.title = 1.4, main = "Intramodular Connectivity", xlab = "genes",
  cex.lab = cex.axis, missing.inds = NULL)
}
\arguments{
\item{adjacency}{Matrix of gene adjacencies}

\item{module.labels}{Optional vector of module assignments for each gene.}

\item{col}{color of bars}

\item{cex.axis}{cex for the axis text}

\item{cex.title}{cex for the title text}

\item{main}{title for the plot}

\item{xlab}{a title for the x axis: see \link[graphics]{title}.}

\item{cex.lab}{cex for the axis label}

\item{missing.inds}{indices for missing genes, needed to preserved order
when comparing plots across datasets.}
}
\description{
Plot the Intramodular Connectivity
}

