% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotKME}
\alias{plotKME}
\title{Plot the Intramodular Module Membership}
\usage{
plotKME(gene.expr, module.labels, heatmap.gradient, cex.axis = 1,
  cex.title = 1.4, main = "Intramodular Module Membership",
  xlab = "genes", cex.lab = cex.axis, missing.inds = NULL)
}
\arguments{
\item{gene.expr}{matrix of gene expression data. Expects columns to be the
genes.}

\item{module.labels}{Optional vector of module assignments for each gene.}

\item{heatmap.gradient}{A vector of colors to use for (or interpolate over)
to render the coexpression.}

\item{cex.axis}{cex for the axis text}

\item{cex.title}{cex for the title text}

\item{main}{title for the plot}

\item{xlab}{label for the x axis}

\item{cex.lab}{cex for the axis label}

\item{missing.inds}{indices for missing genes, needed to preserved order
 when comparing plots across datasets.}
}
\description{
Plot the Intramodular Module Membership
}

