% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotLegend}
\alias{plotLegend}
\title{Plot a vertical gradient legend}
\usage{
plotLegend(gradient, range, cex.axis = 0.8, cex.title = 1,
  main = "Legend", nTicks = 9)
}
\arguments{
\item{gradient}{A vector of colors to create the gradient}

\item{range}{range of values the colors fall over}

\item{cex.axis}{cex for the axis text}

\item{cex.title}{cex for the title text}

\item{main}{title for the legend}

\item{nTicks}{number of tick marks for the axis}

\item{has.missing}{logical; are there missing values on the plot?}
}
\description{
Plot a vertical gradient legend
}

