% Generated by roxygen2 (4.0.2): do not edit by hand
\name{api_inputs}
\alias{api_inputs}
\alias{plotModule}
\title{Plot the topology of a gene coexpression module}
\usage{
plotModule(geneExpression, coexpression, adjacency, moduleAssignments, modules,
  discovery = 1, test = 1, orderSamplesBy = "test",
  orderGenesBy = "discovery", orderModules, plotGeneNames = TRUE,
  plotSampleNames = TRUE, plotModuleNames, main = "Module Topology",
  drawBorders = FALSE, gaxt.line = -0.5, saxt.line = -0.5, maxt.line,
  legend.position = 0.1, legend.tick.size = 0.04, laxt.line = 2.5,
  cex.axis = 0.8, cex.lab = 1, cex.main = 1.2)
}
\arguments{
\item{geneExpression}{the gene expression matrix for the dataset of interest,
or a list of gene expression matrices, one for each dataset. Columns are
expected to be genes, rows samples. See details for expected input types.}

\item{coexpression}{the coexpression matrix for the dataset of interest, or a
list of coexpression matrices, one for each dataset. See details for
expected input types.}

\item{adjacency}{the adjacency matrix for the dataset of interest, or a
list of adjacency matrices, one for each dataset. See details for
expected input types.}

\item{moduleAssignments}{a vector assigning genes to modules, or a list of
such vectors. See details.}

\item{modules}{a vector of modules to apply the function to. See details.}

\item{discovery}{name or index denoting which dataset the module of
interest was discovered in. See details.}

\item{test}{name or index denoting which dataset to apply the function to.
 See details.}

\item{orderSamplesBy}{one of "discovery", "test", or "none". If
"discovery"samples are ordered by their summary expression profile in the
\code{discovery} dataset (see \code{\link{sampleOrder}}). If "test" samples
are ordered by their summary expression profile in the \code{test} dataset.
If "none" no ordering is applied.}

\item{orderGenesBy}{one of "discovery", "test", or "none". If "discovery"
genes are ordered by intramodular connectivity in the \code{discovery}
dataset. If "test" genes are orderd by intramodular connectivity in the
\code{test} dataset. If "none" no ordering is applied.}

\item{orderModules}{logical; if \code{TRUE} modules ordered by similarity of
their summary expression profiles. If \code{FALSE} modules are rendered in
the order provided. The default is to order by modules if the gene
expression is provided.}

\item{plotGeneNames}{logical; if \code{TRUE}, plot the gene names below the
heatmap.}

\item{plotSampleNames}{logical; if \code{TRUE} the sample names will be
plotted next to the gene expression heatmap and summary expression profile
plots}

\item{plotModuleNames}{logical; if \code{TRUE}, plot the module names below
the heatmap. By default, module names are only plotted if multiple
\code{modules} are provided.}

\item{main}{title for the plot.}

\item{drawBorders}{logical; if \code{TRUE}, borders are drawn around the bars
in \code{plotModuleMembership}, \code{plotConnectivity}, or
\code{plotSummaryExpression}.}

\item{cex.axis}{relative size of the gene and sample names.}

\item{cex.lab}{relative size of the module names and legend titles.}

\item{cex.main}{relative size of the plot titles.}

\item{symmetric}{logical; if \code{TRUE} the coexpression and adjacecny
heatmaps will be plotted as symmetric heatmaps, if \code{FALSE} the will be
plotted as a triangular heatmap.}
}
\description{
Plot the topology of one or more network modules.
}
\details{
The matrices containing the \code{geneExpression}, \code{coexpression}, and
 \code{adjacency} should be \code{\link{bigMatrix}} object(s), but the
 function will also work with regular matrices, files containing the matrix
 data, or file paths to the descriptor files for a 'bigMatrix', in which
 case temporary 'bigMatrix' objects will be created. It is advisable to keep
 large matrices stored as 'bigMatrix' objects, as the conversion process
 involves writing out the matrix as binary data to disk. 'bigMatrix' objects
 have the added advantage of having instant load times in new R sessions, and
 can be accessed and used the same way as regular matrices. See
 \code{\link{bigMatrix}} for more details.

 This function can be used in three ways. First, the \code{geneExpression},
 \code{coexpression}, and \code{adjacency} can be provided as lists of
 matrices, where each element corresponds to a dataset of interest. In this
 case, the function will be applied to the \code{test} dataset for the subset
 of genes corresponding to modules discovered in the \code{discovery} dataset.
 By default, the function is applied on dataset 1, on the modules discovered
 in dataset 1. Secondly, the function also accepts single matrices for each,
 simplifying the input for users with only one dataset of interest. Finally,
 \code{modules} and \code{moduleAssignments} may be omitted, in which case
 the function is applied on all genes present in the \code{geneExpression},
 \code{coexpression}, and \code{adjacency}. This may be useful when
 subsetting each matrix by a custom gene set of interest.
}
\seealso{
\code{\link[=plotTopology]{Functions}} for plotting individual
 components of the module topology plot.
}

