% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotModuleLegend}
\alias{plotModuleLegend}
\title{Plot Module legend}
\usage{
plotModuleLegend(module.labels, module.colors)
}
\arguments{
\item{module.labels}{module labels for the nodes in the corresponding
 coexpression heatmap.}
}
\description{
Render a horizontal bar that demarcates more clearly where nodes fall into
separate modules.
}

