% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotNetworkHeatmap}
\alias{PlotNetworkHeatmap}
\title{Plot network edge weights as a heatmap.}
\usage{
PlotNetworkHeatmap(edge.matrix, network.labels = NULL, xlab = "",
  ylab = "", node.labels = NULL, legend.main = "",
  cluster.legend.main = "", legend.bins = 11, edge.weight.range = c(-1,
  1), heatmap.gradient, cluster.cols, mar = c(5.1, 4.1, 2.1, 2.1),
  gap.width = 2.1, layout.heights = c(0.85, 0.15), ...)
}
\arguments{
\item{edge.matrix}{a square matrix containing the edge weights between each
pair of nodes.}

\item{network.labels}{an optional named vector assigning each node to a
network/cluster.}

\item{xlab}{x-axis heatmap label.}

\item{ylab}{y-axis heatmap label.}

\item{node.labels}{a character vector of labels to print for each row/column
cell.}

\item{legend.main}{title to give to the legend denoting heatmap colors.}

\item{cluster.legend.main}{title to give to the legend denoting cluster
assignment.}

\item{legend.bins}{Number of bins to partition heatmap colors into on the
legend. Colors on the heatmap itself are unaffected by this. Colors
are interpolated over the \code{heatmap.gradient}.}

\item{edge.weight.range}{The range of values edge weights can take. Defaults
to \code{c(-1, 1)}.}

\item{heatmap.gradient}{A vector of colors to use for (or interpolate over)
the heatmap.bins.}

\item{cluster.cols}{an optional vector of colors for denoting each cluster
on a legend.}

\item{mar}{outside margins of the plot}

\item{gap.width}{the margin size in lines between the heatmap, and the
\code{cluster.legend}}

\item{layout.heights}{relative proportion of the plot that the heatmap
takes up compared to the cluster labelling. Useful if your cluster
labels are large. Must sum to 1.}
}
\description{
Plots the edge weights between nodes as a heatmap, optionally drawing boxes
around pre-determined clusters. The \code{xlab} and \code{ylab} arguments are
useful when comparing the edge weights.
}
\note{
It expects nodes to be pre-ordered within the edge weight matrix:
 I.e. clusters should be in continuous blocks. Otherwise, garbage-in,
 garbage-out.
}

