% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{plotProps}
\alias{plotProps}
\title{Get the network properties and order for a plot}
\usage{
plotProps(network, data, moduleAssignments, modules, di, ti, orderNodesBy,
  orderSamplesBy, orderModules, datasetNames, nDatasets, verbose, loadedIdx,
  dataLoaded, networkLoaded)
}
\arguments{
\item{network}{list returned by \code{'processInput'}.}

\item{data}{data returned by \code{'processInput'}.}

\item{moduleAssignments}{list returned by \code{'processInput'}.}

\item{modules}{vector of modules to show on the plot.}

\item{di}{name of the discovery dataset.}

\item{ti}{name of the test dataset.}

\item{orderNodesBy}{vector returned by \code{'processInput'}.}

\item{orderSamplesBy}{vector returned by \code{'processInput'}.}

\item{orderModules}{vector returned by \code{'checkPlotArgs'}.}

\item{datasetNames}{vector returned by \code{'processInput'}.}

\item{nDatasets}{vector returned by \code{'processInput'}.}

\item{verbose}{logical; turn on verbose printing.}

\item{loadedIdx}{index of the currently loaded dataset.}

\item{dataLoaded}{currently loaded data matrix (may be NULL).}

\item{networkLoaded}{currently loaded network matrix.}
}
\description{
Get the network properties and order for a plot
}
\keyword{internal}

