% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotSquareHeatmap}
\alias{plotSquareHeatmap}
\title{Plot a square heatmap}
\usage{
plotSquareHeatmap(values, palette, vlim, mas, na.indices.x = NULL,
  na.indices.y = NULL, na.col = "#bdbdbd")
}
\arguments{
\item{values}{values to plot on the heatmap}

\item{palette}{color palette to interpolate over}

\item{vlim}{range of values to use when mapping values to the \code{palette}.}

\item{mas}{ordered subset of the moduleAssignments vector}

\item{na.indices.x}{indices of missing values on the x axis}

\item{na.indices.y}{indices of missing values on the y axis}

\item{na.col}{color of missing values to plot.}
}
\description{
Plot a square heatmap
}

