% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotStatDensities}
\alias{plotStatDensities}
\title{Plot the distributions for the statistics}
\usage{
plotStatDensities(null, observed)
}
\arguments{
\item{null}{Three dimensional array containing the null distributions for
each of the preservation statistics for each network subset.}

\item{observed}{matrix containing the test statistic for each of the
preservation statistics for each network subset.}
}
\description{
Plot the distributions for the statistics
}

