% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotSubsetLegend}
\alias{PlotSubsetLegend}
\title{Plot Subset legend}
\usage{
PlotSubsetLegend(node.labels)
}
\description{
Render a horizontal bar that demarcates where nodes in the
figure generated by \code{\link{PlotNetworkHeatmap}} fall.
}

