% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotTriangleHeatmap}
\alias{plotTriangleHeatmap}
\title{Plot a symmetric heatmap as a triangle}
\usage{
plotTriangleHeatmap(values, palette, vlim, mas)
}
\arguments{
\item{values}{values to plot on the heatmap}

\item{palette}{color palette to interpolate over}

\item{vlim}{range of values to use when mapping values to the \code{palette}.}

\item{mas}{ordered subset of the moduleAssignments vector}
}
\description{
Plot a symmetric heatmap as a triangle
}

