% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{poke}
\alias{poke}
\title{Poke attached big matrix objects to initialise them}
\usage{
poke(...)
}
\arguments{
\item{...}{a number of big.matrix objects}
}
\description{
This is a magical speed hack. See details
}
\details{
For some reason, one of the datasets i tried running on took an incredibly
long time for some very small network subsets. After playing around it turns
out that the first time any of the Rcpp functions are called, regardless of 
subset size, it was taking forever. This can be solved simply by accessing
the big.matrix objects beforehand.
}

