% Generated by roxygen2 (4.0.2): do not edit by hand
\name{preservationPlot}
\alias{preservationPlot}
\title{Generate preservation plot components}
\usage{
preservationPlot(gene.expr, coexpression, adjacency, moduleLabels, module,
  geneIndices, is.relative = TRUE, cex.axis = 0.8, cex.title = 0.9,
  dpi = 100, cex.lab = 0.6)
}
\arguments{
\item{gene.expr}{gene expression matrix. Expects rows to be samples, and
columns to be genes.}

\item{coexpression}{matrix of pairwise gene coexpression.}

\item{adjacency}{matrix of pairwise gene adjacencies.}

\item{moduleLabels}{vector of labels assigning each gene to a module.}

\item{module}{module to plot preservation for.}

\item{geneIndices}{alternative to \code{module}: specify indices of module
instead}

\item{is.relative}{logical; is the gene expression relative to some other
measure? If \code{TRUE}, a divergent palette is selected when plotting the
expression, otherwise a sequential palette is chosen.}

\item{cex.axis}{\code{cex} for the axis tick labels}

\item{cex.title}{\code{cex} for the titles}

\item{dpi}{resolution for each of the panels.#'}

\item{cex.lab}{\code{cex} for the axis labels}
}
\value{
Outputs files for the following:
\itemize{
\item coexpression heatmap
\item coexpression legend
\item intramodular connectivity
\item intramodular module membership
\item expression heatmap
\item summary expression profile
}
}
\description{
Generate preservation plot components
}
\details{
All matrix arguments will take \code{matrix}, \code{big.matrix}, or
\code{big.matrix} descriptors.
}

