\name{PrintProgress}
\alias{PrintProgress}
\title{Prints the progress of a looping function each time the loop progresses by
at least 1\%.}
\usage{
  PrintProgress(this, total)
}
\arguments{
  \item{this}{the current iteration number.}

  \item{total}{the total number of iterations}
}
\description{
  Prints the progress of a looping function each time the
  loop progresses by at least 1\%.
}

