% Generated by roxygen2 (4.0.2): do not edit by hand
\name{qqperm}
\alias{qqperm}
\title{Visual Sanity Check of pperm}
\usage{
qqperm(permuted, q, tfunc, tail.approx = TRUE, lowerTail = FALSE, ...)
}
\arguments{
\item{permuted}{vector of values making up the empirical distribution.}

\item{q}{vector of quantiles to test.}

\item{tfunc}{a probability function to use to get the expected p-values. This
should correspond to the theortical distribution \code{permuted} was drawn
from.}

\item{tail.approx}{logical; if \code{TRUE}, use the tail approximation
algorithm to estimate extreme p-values (see details).}

\item{lowerTail}{logical; if TRUE (default), probabilities are
\eqn{P[x \le x]} otherwise \eqn{P[X > x]}.}

\item{...}{other arguments to pass to \code{\link[graphics]{plot}}. Note that
providing \code{xlab}, \code{ylab}, \code{xlim}, or \code{ylim} will result
in an error.}
}
\description{
Generate a qqplot of the observed p-values from \code{\link{pperm}}, vs the
p-values we would expect from the theoretical distribution the
\code{permuted} observations were drawn from (given in \code{tfunc}).
}
\note{
For internal testing purposes.
}

