% Generated by roxygen2 (4.0.2): do not edit by hand
\name{RangeSubset}
\alias{BigRange}
\alias{RangeSubset}
\title{Get the range of values in a big.matrix column subset}
\usage{
RangeSubset(pDat, subsetIndices)

BigRange(pDat)
}
\arguments{
\item{pDat}{SEXP container for the pointer to the data matrix to be scaled.}

\item{subsetIndices}{indices of the network subset of interest in
  \code{pDat}.}

\item{pDat}{SEXP container for the pointer to the data matrix to be scaled.}

\item{subsetIndices}{indices of the network subset of interest in
  \code{pDat}.}
}
\description{
Get the range of values in a big.matrix column subset

Get the range of values in a big.matrix
}

