% Generated by roxygen2 (4.0.1): do not edit by hand
\name{requiredPower}
\alias{requiredPower}
\title{How many permutations do I need}
\usage{
requiredPower(alpha)
}
\arguments{
\item{alpha}{desired significance threshold}
}
\value{
The minimum number of permutations required to detect any significant
 associations at the provided \code{alpha}
}
\description{
How many permutations do you need to run to be able to detect significance at
a given threshold \code{alpha}?
}

