% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderSamples.R
\name{sampleOrderInternal}
\alias{sampleOrderInternal}
\title{Internal use function for calculating sample order}
\usage{
sampleOrderInternal(props, verbose, na.rm)
}
\arguments{
\item{props}{network properties to calculate order from.}

\item{verbose}{logical; is verbose printing turned on?}

\item{na.rm}{logical; remove missing nodes?}
}
\value{
list structure of ordered nodes.
}
\description{
Used by plotting functions: assumes user input has been sanitised already
}

