% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Scale}
\alias{Scale}
\title{Scale a matrix by its rows}
\usage{
Scale(pDat, spDat)
}
\arguments{
\item{pDat}{SEXP container for the pointer to the data matrix to be scaled.}

\item{spDat}{SEXP container for the pointer to the pre-initialised
\code{\link[bigmemory]{big.matrix}} that the scaled version of \code{pDat}
will be stored in.}
}
\description{
Scale a matrix by its rows
}

