% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Scale}
\alias{Scale}
\title{Scale data across all nodes}
\usage{
Scale(data)
}
\arguments{
\item{data}{matrix to scale.}
}
\value{
A scaled data matrix.
}
\description{
Each node is centered by its mean and scaled by it standard deviation.
}

