% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scaleBigMatrix}
\alias{scaleBigMatrix}
\title{Scale and Center the rows of a 'big.matrix'}
\usage{
scaleBigMatrix(x, obj.dir)
}
\arguments{
\item{x}{the matrix to scale.}

\item{obj.dir}{the directory to store the object in}
}
\value{
A descriptor to the scaled version of x.
}
\description{
Create a new scaled \code{\link[bigmemory]{big.matrix}}.
}
\note{
Unlike \code{\link[base]{scale}} this function applies the scale function to
 the rows of the matrix instead of the columns.

 This function has only been implemented for
 \code{\link[bigmemory]{big.matrix}} objects of \code{type = "double"}.

 To apply the scale function in place without creating a new results matrix,
 simply call \code{Scale(x@address, x@address)}.
}
\seealso{
\code{\link[base]{scale}} \code{\link[bigmemory]{big.matrix}}
}

