% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SetDiag}
\alias{SetDiag}
\title{big.matrix diagonals}
\usage{
SetDiag(pBigMat, value)
}
\arguments{
\item{pBigMat}{a \code{numeric} \code{\link[bigmemory]{big.matrix}}}

\item{value}{either a single \code{numeric} value or \code{numeric} vector
of length equal to the current diagonal.}
}
\description{
big.matrix diagonals
}

