% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-defs.R
\name{simplify_param}
\alias{simplify_param}
\title{Template parameters}
\arguments{
\item{simplify}{logical; if \code{TRUE}, simplify the structure of the output
list if possible (see Return Value).}

\item{nCores}{number of cores to parallelise the calculation of network 
properties over. Ignored if the user has already registered a parallel 
backend.}
}
\description{
Template parameters to be imported into other function documentation. This 
is not intended to be a stand-alone help file.

Template parameters to be imported into other function documentation. This 
is not intended to be a stand-alone help file.
}

