% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simplifyList}
\alias{simplifyList}
\title{Remove unnecessary list structure}
\usage{
simplifyList(l, depth)
}
\arguments{
\item{l}{a nested list}

\item{depth}{depth to traverse to}
}
\value{
a list
}
\description{
Removes entries that are \code{NULL} are extracts element if the length is 1.
}

