% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simplifyList2}
\alias{simplifyList2}
\title{Remove unnecessary list structure at depth = 2}
\usage{
simplifyList2(l)
}
\arguments{
\item{l}{a nested list}
}
\value{
a list
}
\description{
Removes entries that are \code{NULL} are extracts element if the length is 1.
}

