% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sortModuleNames}
\alias{sortModuleNames}
\title{Get a sorted list of module names}
\usage{
sortModuleNames(modules)
}
\arguments{
\item{modules}{a vector of module labels to sort}
}
\value{
a sorted vector
}
\description{
If module labels are numeric, sorts numerically, otherwise sorts 
alphabetically.
}
\keyword{internal}

