% Generated by roxygen2 (4.0.1): do not edit by hand
\name{subsetTestStats}
\alias{subsetTestStats}
\title{Between-dataset network statistics.}
\usage{
subsetTestStats(discProps, testProps)
}
\arguments{
\item{discProps,testProps}{properties of the network subset in the each of
the \emph{discovery} and \emph{test} datasets, respectively, as calculated
by \code{\link{subsetProps}}.}
}
\value{
A vector of test statistics.
}
\description{
Calculate network subset replication test statistics from their network
properties (as calculated by \code{\link{subsetProps}}).
}
\details{
The returned test statistics indicate the replication/preservation of each
network property for a network subset in the two datasets provided in the
\code{discProps} and \code{testProps} arguments.

Evaluating the significance of these test statistics requires generation of a
null distribution for each, by randomly sampling network subsets of the same
size in the \emph{test} dataset, calculating their replication, and
evaluating how extreme the observed test statistic is, in comparison to test
statistics drawn from random permutation. This procedure is performed by the
main function of this package, \code{\link{netRep}}.
}
\references{
Langfelder, P., Luo, R., Oldham, M. C. & Horvath, S. \emph{Is my network
module preserved and reproducible?} PLoS Comput. Biol. \strong{7}, e1001057
(2011).
}
\seealso{
\code{\link[=subsetProps]{Network subset topology}}
  \code{\link{netRep}}
}

