% Generated by roxygen2 (4.0.1): do not edit by hand
\name{vCat}
\alias{vCat}
\title{Verbose Concatenate and Print}
\usage{
vCat(verbose, ...)
}
\arguments{
\item{verbose}{logical. If \code{TRUE}, passes the rest of the arguments to
\code{\link{cat}}}

\item{...}{arguments to pass to \code{\link{cat}}}
}
\description{
Concatenate and output the objects only if the \code{verbose} flag is set
to \code{TRUE}.
}

