\name{VerbosePrint}
\alias{VerbosePrint}
\title{Prints the message only if the verbose flag is set.}
\usage{
  VerbosePrint(..., verbose)
}
\arguments{
  \item{...}{message to print}

  \item{verbose}{Logical; prints message if true}
}
\description{
  Prints the message only if the verbose flag is set.
}

