% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-user-input.R
\name{verifyDatasetOrder}
\alias{verifyDatasetOrder}
\title{Verify a 'list' input ordering}
\usage{
verifyDatasetOrder(tocheck, errname, dataNames, nDatasets)
}
\arguments{
\item{tocheck}{list of data to check.}

\item{errname}{name to print in error messages.}

\item{dataNames}{names of the datasets.}

\item{nDatasets}{number of datasets.}
}
\value{
ordered 'tocheck' by dataset.
}
\description{
Check and order an input list:
}
\keyword{internal}

