% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-functions.R
\name{addGradientLegend}
\alias{addGradientLegend}
\title{Plot a color palette legend}
\usage{
addGradientLegend(palette, palette.vlim, legend.vlim, horizontal, main, xlim,
  ylim, tick.size = 0.04, axis.line = 3, border.width = 2, srt)
}
\arguments{
\item{palette}{color palette.}

\item{palette.vlim}{limits of the values mapping to the extremities of the 
color palette.}

\item{legend.vlim}{limits of the values to display on the legend}

\item{horizontal}{logical; if \code{TRUE} the legend is plotted horizontally,
otherwise vertically.}

\item{main}{title of the legend.}

\item{xlim}{xlim relative to the plotting region of the rest of the plot.}

\item{ylim}{ylim relative to the plotting region of the rest of the plot.}

\item{tick.size}{size of the legend axis ticks relative to the size of the 
plot window.}

\item{axis.line}{distance from the axis to render the axis labels as a 
multiple of \code{tick.size}.}

\item{border.width}{line width for borders.}

\item{srt}{angle of text labels}
}
\description{
Add a legend to a plot window.
}

