% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-functions.R
\name{addGradientLegend}
\alias{addGradientLegend}
\title{Plot a color palette legend}
\usage{
addGradientLegend(palette, legend.vlim, horizontal, main, xlim, ylim,
  tck = 0.04, axis.line = 3, legend.main.line = 1, lwd = 2, srt)
}
\arguments{
\item{palette}{color palette.}

\item{legend.vlim}{limits of the values to display on the legend}

\item{horizontal}{logical; if \code{TRUE} the legend is plotted horizontally,
otherwise vertically.}

\item{main}{title of the legend.}

\item{xlim}{xlim relative to the plotting region of the rest of the plot.}

\item{ylim}{ylim relative to the plotting region of the rest of the plot.}

\item{tck}{size of the legend axis ticks relative to the size of the 
plot window.}

\item{axis.line}{distance from the axis to render the axis labels as a 
multiple of \code{tck}.}

\item{legend.main.line}{distance from the plot to render the legend title.}

\item{lwd}{line width for borders.}

\item{srt}{angle of text labels}
}
\description{
Add a legend to a plot window.
}
\keyword{internal}

