% Generated by roxygen2 (4.0.2): do not edit by hand
\name{adj_param}
\alias{adj_param}
\title{Template parameters}
\arguments{
\item{adj}{a \code{\link[=bigMatrix-class]{bigMatrix}} containing the
  pairwise gene adjacencies.}
}
\description{
Template parameters to be imported into other function documentation. This
is not intended to be a stand-alone help file.
}

