% Generated by roxygen2 (4.0.2): do not edit by hand
\name{AdjProps}
\alias{AdjProps}
\title{Calculate Mean Adjacency and Intramodular Connectivity}
\usage{
AdjProps(pAdjacency, subsetIndices)
}
\arguments{
\item{pAdjacency}{SEXP container for the pointer to the adjacency matrix.}

\item{subsetIndices}{indices of the network subset of interest.}
}
\value{
A List containing:
  \enumerate{
    \item{\emph{kIM}:}{The weighted within-subset degree for each node.}
    \item{\emph{meanAdj}:}{The mean absolute edge weight of the network subset.}
  }
}
\description{
Calculate Mean Adjacency and Intramodular Connectivity
}

