% Generated by roxygen2 (4.0.2): do not edit by hand
\name{api_inputs}
\alias{api_inputs}
\alias{networkProperties}
\title{Get the properties of a network module}
\usage{
networkProperties(geneExpression = NULL, coexpression, adjacency,
  moduleAssignments, modules, discovery = 1, test = 1)
}
\arguments{
\item{geneExpression}{a \code{\link{bigMatrix}} object, or a list of
\code{bigMatrix} objects, see details.}

\item{coexpression}{a \code{\link{bigMatrix}} object, or a list of
\code{bigMatrix} objects, see details.}

\item{adjacency}{a \code{\link{bigMatrix}} object, or a list of
\code{bigMatrix} objects, see details.}

\item{moduleAssignments}{a vector assigning genes to modules, or a list of
such vectors. See details.}

\item{modules}{a vector of modules to apply the function to. See details.}

\item{discovery}{name or index denoting which dataset the module of
interest was discovered in. See details.}

\item{test}{name or index denoting which dataset to apply the function to.
 See details.}
}
\value{
A list of network properties for each module of interest.
}
\description{
Calculate the network properties that compose the module preservation
statistics for one or more modules.
}
\details{
If any of the \code{geneExpression}, \code{coexpression}, \code{adjacency},
or \code{moduleAssignments} arguments are list objects, then they all must
be. In this case, each list element corresponds to a dataset that the
respective network components have been calculated in. Providing the
\code{geneExpression} is optional, but limits the scope of the analysis or
function output. The \code{discovery} argument denotes which dataset the
\code{modules} of interest were discovered in, and the \code{test} argument
controls which dataset the function is calculated on. Alternatively, the
\code{geneExpression}, \code{coexpression}, \code{adjacency}, and
\code{moduleAssignments} can be provided only for the discovery dataset.
}

