% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{big.matrix-extensions}
\alias{big.matrix-extensions}
\alias{diag,big.matrix-method}
\alias{diag<-,big.matrix-method}
\title{big.matrix diagonals}
\usage{
\S4method{diag}{big.matrix}(x) <- value

\S4method{diag}{big.matrix}(x)
}
\arguments{
\item{x}{a matrix, vector or 1D array, or missing.}

\item{value}{either a single value or a vector of length equal to that
    of the current diagonal.  Should be of a mode which can be coerced
    to that of \code{x}.}
}
\description{
Extract or replace the diagonal of a \code{\link[bigmemory]{big.matrix}}
}
\section{Additional Notes}{

 Unlike \code{\link[base]{diag}} for \code{\link[base]{matrix}} objects, you
 cannot use \code{diag} to construct a new \code{big.matrix}.
}

