% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{as.big.matrix,bigMatrix-method}
\alias{as.big.matrix,bigMatrix-method}
\alias{as.matrix,bigMatrix-method}
\alias{write.bigMatrix}
\title{Writing out data, or converting from, a 'bigMatrix'.}
\usage{
\S4method{as.big.matrix}{bigMatrix}(x)

\S4method{as.matrix}{bigMatrix}(x)

write.bigMatrix(x, file, ...)
}
\arguments{
\item{x}{an object of type 'bigMatrix'.}

\item{file}{file to write out the data from \code{x} into.}

\item{...}{additional arguments to pass to
 \code{\link[bigmemory]{write.big.matrix}}.}
}
\description{
Functions for converting 'bigMatrix' objects to other classes.
}
\details{
\code{as.big.matrix} converts to a \code{\link[bigmemory]{big.matrix}}
object.
\code{as.matrix} converts to a \code{\link[base]{matrix}} object.
\code{write.bigMatrix} will write out the data into a regularly structure file
(see \code{\link[utils]{write.table}})
}

