% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigMatrix.R
\name{load.bigMatrix}
\alias{load.bigMatrix}
\title{Load a 'bigMatrix' (deprecated)}
\usage{
load.bigMatrix(backingfile)
}
\description{
The \code{'bigMatrix'} class is no longer implemented in the \code{NetRep}
package: the shared memory approach was incompatabile with high performance
compute clusters, so the parallel permutation procedure has been translated
into C++ code (which is also much faster). For those who had saved their 
data in the \code{'bigMatrix'} format, this function will convert these 
objects to the similar \code{'\link[bigmemory]{big.matrix}'} class
from the \code{\link{bigmemory}} package.
}

