% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CheckFinite}
\alias{CheckFinite}
\title{Check whether there are any non-finite values in a matrix}
\usage{
CheckFinite(matPtr)
}
\arguments{
\item{matPtr}{matrix to check.}
}
\value{
Throws an error if any \code{NA}, \code{NaN}, \code{Inf}, or \code{-Inf}
 values are found, otherwise returns silently.
}
\description{
The C++ functions will not work with NA values, and the calculation of the
summary profile will take a long time to run before crashing.
}
\keyword{internal}

