% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkFinite}
\alias{checkFinite}
\title{Check if all entries of a `big.matrix` are Finite}
\usage{
checkFinite(x)
}
\arguments{
\item{x}{a \code{\link[bigmemory]{big.matrix}}}
}
\description{
If there are non-finite entires (\code{NA}, \code{NaN}, \code{-Inf},
\code{Inf}), throw an exception.
}

