% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-user-input.R
\name{checkPlotArgs}
\alias{checkPlotArgs}
\title{Validate plot function arguments}
\usage{
checkPlotArgs(orderModules, plotNodeNames, plotSampleNames, plotModuleNames,
  main, drawBorders, lwd, naxt.line, saxt.line, maxt.line, xaxt.line, yaxt.line,
  laxt.line, xaxt.tck, yaxt.tck, laxt.tck, xlab.line, ylab.line, main.line,
  plotLegend, legend.position, legend.main, legend.main.line, symmetric,
  horizontal, dataCols, dataRange, corCols, corRange, netCols, netRange,
  degreeCol, contribCols, summaryCols, naCol, dryRun)
}
\arguments{
\item{orderModules}{user input for the corresponding argument in the plot functions.}

\item{plotNodeNames}{user input for the corresponding argument in the plot functions.}

\item{plotSampleNames}{user input for the corresponding argument in the plot functions.}

\item{plotModuleNames}{user input for the corresponding argument in the plot functions.}

\item{main}{user input for the corresponding argument in the plot functions.}

\item{drawBorders}{user input for the corresponding argument in the plot functions.}

\item{lwd}{user input for the corresponding argument in the plot functions.}

\item{naxt.line}{user input for the corresponding argument in the plot functions.}

\item{saxt.line}{user input for the corresponding argument in the plot functions.}

\item{maxt.line}{user input for the corresponding argument in the plot functions.}

\item{xaxt.line}{user input for the corresponding argument in the plot functions.}

\item{yaxt.line}{user input for the corresponding argument in the plot functions.}

\item{laxt.line}{user input for the corresponding argument in the plot functions.}

\item{xaxt.tck}{user input for the corresponding argument in the plot functions.}

\item{yaxt.tck}{user input for the corresponding argument in the plot functions.}

\item{laxt.tck}{user input for the corresponding argument in the plot functions.}

\item{xlab.line}{user input for the corresponding argument in the plot functions.}

\item{ylab.line}{user input for the corresponding argument in the plot functions.}

\item{main.line}{user input for the corresponding argument in the plot functions.}

\item{plotLegend}{user input for the corresponding argument in the plot functions.}

\item{legend.position}{user input for the corresponding argument in the plot functions.}

\item{legend.main}{user input for the corresponding argument in the plot functions.}

\item{legend.main.line}{input for the corresponding argument in the plot functions.}

\item{symmetric}{user input for the corresponding argument in the plot functions.}

\item{horizontal}{user input for the corresponding argument in the plot functions.}

\item{dataCols}{user input for the corresponding argument in the plot functions.}

\item{dataRange}{user input for the corresponding argument in the plot functions.}

\item{corCols}{user input for the corresponding argument in the plot functions.}

\item{corRange}{user input for the corresponding argument in the plot functions.}

\item{netCols}{user input for the corresponding argument in the plot functions.}

\item{netRange}{user input for the corresponding argument in the plot functions.}

\item{degreeCol}{user input for the corresponding argument in the plot functions.}

\item{contribCols}{user input for the corresponding argument in the plot functions.}

\item{summaryCols}{user input for the corresponding argument in the plot functions.}

\item{naCol}{user input for the corresponding argument in the plot functions.}

\item{dryRun}{user input for the corresponding argument in the plot functions.}
}
\description{
Simple typechecking for the extensive plot arguments
}

