% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-user-input.R
\name{checkPlotArgs}
\alias{checkPlotArgs}
\title{Validate plot function arguments}
\usage{
checkPlotArgs(orderModules, plotNodeNames, plotSampleNames, plotModuleNames,
  main, drawBorders, border.width, naxt.line, saxt.line, maxt.line,
  legend.tick.size, laxt.line, plotLegend, legend.position, legend.main,
  palette, symmetric, horizontal)
}
\arguments{
\item{orderModules}{user input for the corresponding argument in the plot functions.}

\item{plotSampleNames}{user input for the corresponding argument in the plot functions.}

\item{plotModuleNames}{user input for the corresponding argument in the plot functions.}

\item{main}{user input for the corresponding argument in the plot functions.}

\item{drawBorders}{user input for the corresponding argument in the plot functions.}

\item{border.width}{user input for the corresponding argument in the plot functions.}

\item{naxt.line}{user input for the corresponding argument in the plot functions.}

\item{saxt.line}{user input for the corresponding argument in the plot functions.}

\item{maxt.line}{user input for the corresponding argument in the plot functions.}

\item{legend.tick.size}{user input for the corresponding argument in the plot functions.}

\item{laxt.line}{user input for the corresponding argument in the plot functions.}

\item{plotLegend}{user input for the corresponding argument in the plot functions.}

\item{legend.position}{user input for the corresponding argument in the plot functions.}

\item{legend.main}{user input for the corresponding argument in the plot functions.}

\item{palette}{user input for the corresponding argument in the plot functions.}

\item{symmetric}{user input for the corresponding argument in the plot functions.}

\item{horizontal}{user input for the corresponding argument in the plot functions.}

\item{plotNodesNames}{user input for the corresponding argument in the plot functions.}
}
\description{
Simple typechecking for the extensive plot arguments
}

