% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-user-input.R
\name{checkSets}
\alias{checkSets}
\title{Check input data for consistency}
\usage{
checkSets(data, correlation, network, moduleAssignments, discovery, test)
}
\arguments{
\item{data}{a list of numeric matrices. Each entry of the list corresponds to
a dataset and contains the data used to infer the interaction network
between variables (e.g. genes). Expects matrix columns to correspond to
variables and matrix rows to correspond to samples.}

\item{correlation}{a list of matrices. Each entry of the list corresponds to a 
dataset and contains an \eqn{n * n} matrix of the correlation between 
each pair of variables in the dataset.}

\item{network}{a list of matrices. Each entry of the list corresponds to a 
dataset and contains an \eqn{n * n} matrix of the network edge weights 
between each pair of variables in the dataset.}

\item{moduleAssignments}{a vector containing the module each variable belongs
to in the discovery dataset. If there are multiple discovery datasets 
then this argument should be a list of such vectors.}

\item{discovery}{name or index denoting which dataset the module of interest
was discovered in.}

\item{test}{name or index denoting which dataset to apply the function to.}
}
\description{
Make sure all objects are the right class, have the same dimensions, and
are ordered the same way across datasets.
}

