% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-user-input.R
\name{checkSets}
\alias{checkSets}
\title{Check input data for consistency}
\usage{
checkSets(geneExpression, coexpression, adjacency, moduleAssignments, discovery,
  test)
}
\arguments{
\item{geneExpression}{a list of \code{bigMatrix} objects}

\item{coexpression}{a list of \code{bigMatrix} objects}

\item{adjacency}{a list of \code{bigMatrix} objects}

\item{moduleAssignments}{a list of either NULL, or vectors assigning genes to
modules in the \code{discovery} datasets.}

\item{discovery}{name or index denoting which dataset the module of
interest was discovered in. See details.}

\item{test}{name or index denoting which dataset to apply the function to.
See details.}
}
\description{
Make sure all objects are the right class, have the same dimensions, and
are ordered the same way across datasets.
}

