% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{col2hex}
\alias{col2hex}
\title{Get the hexidecimal value for a named color}
\usage{
col2hex(name)
}
\arguments{
\item{name}{the color name}
}
\value{
the hex value for that color
}
\description{
Get the hexidecimal value for a named color
}

