% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi-machine.R
\name{combineAnalyses}
\alias{combineAnalyses}
\title{Multi machine analysis}
\usage{
combineAnalyses(pres1, pres2)
}
\arguments{
\item{pres1, pres2}{lists returned by \code{\link{modulePreservation}}.}
}
\value{
A nested list containing the same elements as 
 \code{\link{modulePreservation}}.
}
\description{
This function takes the output from multiple runs of 
\code{\link{modulePreservation}}, combines their results, and returns a new
set of permutation test P-values. This is useful for parallelising 
calculations across multiple machines.
}
\examples{
\dontrun{
data("NetRep")

# Convert them to the 'bigMatrix' format:
discovery_data <- as.bigMatrix(discovery_data)
discovery_correlation <- as.bigMatrix(discovery_correlation)
discovery_network <- as.bigMatrix(discovery_network)
test_data <- as.bigMatrix(test_data)
test_correlation <- as.bigMatrix(test_correlation)
test_network <- as.bigMatrix(test_network)

# Set up input lists for each input matrix type across datasets:
data_list <- list(discovery=discovery_data, test=test_data)
correlation_list <- list(discovery=discovery_correlation, test=test_correlation)
network_list <- list(discovery=discovery_network, test=test_network)
labels_list <- list(discovery=module_labels)

# Assess module preservation on two different machines to spread computational
# burden.
pres1 <- modulePreservation(
 data=data_list, correlation=correlation_list, network=network_list,
 moduleAssignments=labels_list, nPerm=500, discovery="discovery", 
 test="test"
)

pres2 <- modulePreservation(
 data=data_list, correlation=correlation_list, network=network_list,
 moduleAssignments=labels_list, nPerm=500, discovery="discovery", 
 test="test"
)

combined <- combineAnalyses(pres1, pres2)

}

}

