% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingency.R
\name{contingencyTable}
\alias{contingencyTable}
\title{Calculate a contigency table of module overlap between datasets}
\usage{
contingencyTable(moduleAssignments, modules, network, di, ti)
}
\arguments{
\item{moduleAssignments}{processed user input for the 'moduleAssignments'
argument (see \link{processInput})}

\item{modules}{user processed input for the 'modules' argument.}

\item{network}{user processd input for the 'network' argument.}

\item{di}{the discovery dataset to use}

\item{ti}{the test dataset to use}
}
\value{
a list containing a contigency table,
}
\description{
Calculate a contigency table of module overlap between datasets
}

