% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CorStats}
\alias{CorStats}
\title{Calculate the correlation structure based statistics}
\usage{
CorStats(pCorD, discIndices, pCorT, testIndices)
}
\arguments{
\item{pCorD, pCorT}{SEXP containers for the pointers to the correlation 
structure matrices for the \emph{discovery} and \emph{test} networks.}

\item{discIndices, testIndices}{indices of the network subset in
the \emph{discovery} and \emph{test} networks respectively.}
}
\value{
A vector containing:
  \enumerate{
    \item{\emph{cor.discovery}:}{
      A flattened vector of the module's correlation structure in the 
      \emph{discovery} dataset.
    }
    \item{\emph{cor.test}:}{
      A flattened vector of the module's correlation structure in the 
      \emph{test} dataset.
    }
    \item{\emph{corDensity}:}{
      The mean sign-aware correlation structure density of the network module.
    }
  }
}
\description{
Both of the "concordance of correlation structure" and "density of 
correlation structure" are calculated using all pairwise 
correlation coefficients between nodesin both the \emph{discovery} and 
\emph{test} datasets. For the other module preservation statistics we can
store components of each statistic calculated in each dataset separately,
and save time by only calculating them in the discovery dataset once. 
This would have a substantial memory overhead for the correlation structure
based statistics, so we don't use this strategy.
}
\references{
\enumerate{
    \item{
      Langfelder, P., Luo, R., Oldham, M. C. & Horvath, S. \emph{Is my 
      network module preserved and reproducible?} PLoS Comput. Biol. 
      \strong{7}, e1001057 (2011). 
    }
 }
}

