% Generated by roxygen2 (4.0.2): do not edit by hand
\name{DataProps}
\alias{DataProps}
\title{Network subset eigenvector and proportion of variance explained in C++}
\usage{
DataProps(pDat, subsetIndices)
}
\arguments{
\item{pDat}{SEXP container for the pointer to a scaled version of the
data matrix used to construct the network.}

\item{subsetIndices}{indices of the network subset of interest in
  \code{pDat}.}
}
\value{
A list containing:
 \enumerate{
  \item{\emph{"SEP"}:}{
    The Summary Expression Profile of each node (see details).
  }
  \item{\emph{"MM"}:}{
    The Module Membership of each node (see details).
  }
  \item{\emph{"pve"}:}{
    The proportion of the variance explained by the subset's summary
    expression profile (see details).
  }
 }
}
\description{
Network subset eigenvector and proportion of variance explained in C++
}
\details{
First, a summar expression profile (SEP) is calculated for the
 module from the underlying gene expression data. This corresponds to the
 first eigenvector of a principal component analysis \emph{(1)}.

 The orientation of the eigenvector is modified so that its sign is in the
 same direction as the gene expression (on average).

 The Module Membership (MM) is thus quantified as the correlation between each
 gene in the module and the summary expression profile.

 The proportion of variance in the module's gene expression data explained
 by the summary expression profile (pve) is quantified as the average square
of the Module Membership \emph{(1)}.
}
\references{
\enumerate{
    \item{
      Langfelder, P., Luo, R., Oldham, M. C. & Horvath, S. \emph{Is my
      network module preserved and reproducible?} PLoS Comput. Biol.
      \strong{7}, e1001057 (2011).
    }
 }
}

