% Generated by roxygen2 (4.0.1): do not edit by hand
\name{DataSummary}
\alias{DataSummary}
\title{Network subset eigenvector and proportion of variance explained in C++}
\usage{
DataSummary(pDat, pScaledDat, subsetIndices)
}
\arguments{
\item{pDat}{SEXP container for the pointer to the data matrix used in
network construction.}

\item{pScaledDat}{SEXP container for the pointer to a scaled version of the
data matrix used to construct the network.}

\item{subsetIndices}{indices of the network subset of interest in
  \code{pDat}.}
}
\value{
A list whose first element is the first eigenvector of the singular value
 decomposition for the network subset, and whose second element is the
 proportion of the variance in the corresponding subset of \code{pDat} the
 eigenvector explains.
}
\description{
Network subset eigenvector and proportion of variance explained in C++
}
\details{
The sign of the returned eigenvector is modified to match the average of
 \code{pDat}. This is to match the behaviour of
 \code{\link[WGCNA]{moduleEigengenes}} in the \code{\link{WGCNA}} package.

 The two returned properties are bundled together into one function because
 the calculation of the proportion of variance requires much of the same
 underlying intermediate calculations that obtaining the first eigenvector
 requires.
}

