% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-user-input.R
\name{dynamicMatLoad}
\alias{dynamicMatLoad}
\title{Dynamically detect load matrices}
\usage{
dynamicMatLoad(object)
}
\arguments{
\item{object}{user input object}
}
\value{
A matirx or an error.
}
\description{
Dynamically detect load matrices
}

