% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dynamicMatLoad}
\alias{dynamicMatLoad}
\title{Dynamically detect and load a bigMatrix object depending on input type}
\usage{
dynamicMatLoad(object, ...)
}
\arguments{
\item{object}{user input object}

\item{...}{additional arguments to pass to read.bigMatrix}
}
\value{
A 'bigMatrix' object or error.
}
\description{
Dynamically detect and load a bigMatrix object depending on input type
}

