% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-user-input.R
\name{dynamicMatLoad}
\alias{dynamicMatLoad}
\title{Dynamically detect and load a bigMatrix object depending on input type}
\usage{
dynamicMatLoad(object, ...)
}
\arguments{
\item{object}{user input object}

\item{...}{additional arguments to pass to read.bigMatrix}
}
\value{
A 'bigMatrix' object or error.
}
\description{
Dynamically detect and load a bigMatrix object depending on input type
}

