% Generated by roxygen2 (4.0.2): do not edit by hand
\name{forceLim}
\alias{forceLim}
\title{Get the plot limits to set for the desired plot window}
\usage{
forceLim(dlim)
}
\arguments{
\item{dlim}{the limits you want to set}
}
\value{
the min and max to set for the desired plot limits
}
\description{
\code{plot} manually adds an extra region to the plot on top of the given
'xlim' and 'ylim', amounting to 4% of the plot region in either direction.
This function tells you what limits to set so that the boundaries of the plot
are precisely the min and max you want.
}

