% Generated by roxygen2 (4.0.2): do not edit by hand
\name{formatExclude}
\alias{formatExclude}
\title{Format excludeModules list}
\usage{
formatExclude(excludeModules, discovery, nDatasets, datasetNames)
}
\arguments{
\item{discovery}{vector of discovery networks provided by the user}

\item{nDatasets}{number of datasets}

\item{datasetNames}{names of the datasets}

\item{input}{input provided by the user}
}
\value{
List structure to match the rest of the input data
}
\description{
Format excludeModules list
}

