% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-user-input.R
\name{formatInclude}
\alias{formatInclude}
\title{Format the list of modules to include.}
\usage{
formatInclude(include, discovery, nDatasets, datasetNames)
}
\arguments{
\item{include}{input provided by the user}

\item{discovery}{vector of discovery networks provided by the user}

\item{nDatasets}{number of datasets}

\item{datasetNames}{names of the datasets}
}
\value{
List structure to match the rest of the input data
}
\description{
Format the list of modules to include.
}

