% Generated by roxygen2 (4.0.2): do not edit by hand
\name{formatInclude}
\alias{formatInclude}
\title{Format includeModules list}
\usage{
formatInclude(includeModules, discovery, nDatasets, datasetNames)
}
\arguments{
\item{includeModules}{input provided by the user}

\item{discovery}{vector of discovery networks provided by the user}

\item{nDatasets}{number of datasets}

\item{datasetNames}{names of the datasets}
}
\value{
List structure to match the rest of the input data
}
\description{
Format includeModules list
}

