% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getColFromPalette}
\alias{getColFromPalette}
\title{Map a variable to a color gradient}
\usage{
getColFromPalette(values, palette, vlim)
}
\arguments{
\item{values}{a vector of values to retrieve the colors for.}

\item{palette}{a vector of colors}

\item{vlim}{limits of the values}
}
\value{
The color for a given value.
}
\description{
For a color gradient mapped onto a scale of values (\code{vlim}), retrieve
the color for any given \code{value}
}

