% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{getModuleBreaks}
\alias{getModuleBreaks}
\title{Get module break points on the x-axis}
\usage{
getModuleBreaks(mas)
}
\arguments{
\item{mas}{ordered subset of the moduleAssignments vector}
}
\value{
a vector of positions on the x-axis where one module begins and another ends
}
\description{
Get module break points on the x-axis
}

