% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigMatrix-methods.R
\name{is.bigMatrix}
\alias{is.bigMatrix}
\title{Checks if an object is a bigMatrix}
\usage{
is.bigMatrix(x)
}
\arguments{
\item{x}{object to check}
}
\value{
\code{TRUE} if 'x' is a 'bigMatrix', \code{FALSE} otherwise.
}
\description{
Checks if an object is a bigMatrix
}
\seealso{
\code{\link[=bigMatrix-class]{bigMatrix}}
}

