% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.equal}
\alias{is.equal}
\title{Floating Point Comparison}
\usage{
is.equal(vector, value)
}
\arguments{
\item{vector}{a numeric vector of doubles}

\item{value}{a double}
}
\value{
logical; \code{TRUE} where an element is equal to \code{value}, 
  \code{NA} where a comparison is made with an \code{NA}, and \code{FALSE}
  otherwise.
}
\description{
Tests elements in a vector for equality to a specified floating point value.
}

