% Generated by roxygen2 (4.0.1): do not edit by hand
\name{kIM}
\alias{kIM}
\title{Node Connectivity (weighted degree)}
\usage{
kIM(adjacency, subsetIndices, allNodes)
}
\arguments{
\item{adjacency}{Adjacency matrix representation of the network}

\item{subsetIndices}{row/column indices in the provided \code{adjacency} that
correspond to the network subset of interest.}

\item{allNodes}{Logical; if FALSE, the connectivity is calculated based only
on nodes in the module, otherwise a module's connectivity to all nodes is
calculated.}
}
\value{
The connectivity (weighted degree) for each node in the network subset
}
\description{
If \code{allNodes} is \code{FALSE}, then the \emph{intramodular connectivity}
(kIM) is calculated: this is the sum of the edgeweights for each node in the
network subset, to all other nodes in that network subset.
If instead we specify \code{allNodes = TRUE}, then the
\emph{whole network connectivity} for those nodes is calculated.
}
\references{
Langfelder, Peter; Luo, Rui; Oldham, Michael C.; and Horvath,
   Steve. Is My Network Module Preserved and Reproducible?. PLoS
   Computational Biology, 2011.
}

