% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadIntoRAM}
\alias{loadIntoRAM}
\title{Load a \code{'disk.matrix'} into RAM}
\usage{
loadIntoRAM(x)
}
\arguments{
\item{x}{a \code{'matrix'} or \code{'disk.matrix'}}
}
\value{
a \code{'matrix'}
}
\description{
If \code{x} is a \code{\link{disk.matrix}} load in the matrix data at its
associated file. If \code{x} is already a matrix, return as is.
}
\keyword{internal}

