% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MeanAdj}
\alias{MeanAdj}
\title{C++ implementation of Mean Adjacency}
\usage{
MeanAdj(pAdjacency, subsetIndices)
}
\arguments{
\item{pAdjacency}{SEXP container for the pointer to the adjacency matrix}

\item{subsetIndices}{indices of the subset of the network to calculate
the mean adjacency for.}
}
\value{
A single numeric value.
}
\description{
See the \link[=meanAdj]{wrapper function} for documentation.
}

