% Generated by roxygen2 (4.0.1): do not edit by hand
\name{meanAdj}
\alias{meanAdj}
\title{Calculate the meanAdj statistic of a module in a test network}
\usage{
meanAdj(adjacency, moduleIndices)
}
\arguments{
\item{adjacency}{Adjacency matrix to calculate meanAdj of module in.}

\item{moduleIndices}{Where the nodes in the module are located in the
adjacency network.}
}
\value{
The mean edge weight of the module in the test network.
}
\description{
Calculates the mean edge weight of the module in the test network.
A high preservation score indicates preservation of the module's density.
}
\references{
Langfelder, Peter; Luo, Rui; Oldham, Michael C.; and Horvath,
   Steve. Is My Network Module Preserved and Reproducible?. PLoS
   Computational Biology, 2011.
}

