% Generated by roxygen2 (4.0.1): do not edit by hand
\name{meanAdj}
\alias{meanAdj}
\title{Mean Adjacency of a Network}
\usage{
meanAdj(adjacency, moduleIndices)
}
\arguments{
\item{adjacency}{Adjacency matrix.}

\item{subsetIndices}{indices of the subset of nodes to calculate the mean
adjacency on.}
}
\value{
a single numeric value.
}
\description{
The mean adjacency statistic measures how \emph{dense} a network is. It is
simply the average edge weight of the given network, or network subset.
}
\references{
Langfelder, Peter; Luo, Rui; Oldham, Michael C.; and Horvath,
   Steve. Is My Network Module Preserved and Reproducible?. PLoS
   Computational Biology, 2011.
}

