% Generated by roxygen2 (4.0.1): do not edit by hand
\name{meanAdjR}
\alias{meanAdjR}
\title{Mean Adjacency of a Network, R Implementation.}
\usage{
meanAdjR(adjacency, subsetIndices)
}
\arguments{
\item{adjacency}{Adjacency matrix.}

\item{subsetIndices}{row/column indices in the provided \code{adjacency} that
correspond to the network subset of interest.}
}
\value{
a single numeric value.
}
\description{
Used to unit test the C++ functionality of \code{\link{meanAdj}}.
}

