% Generated by roxygen2 (4.0.1): do not edit by hand
\name{NetProps}
\alias{NetProps}
\title{Calculate Network Properties}
\usage{
NetProps(pAdjacency, subsetIndices)
}
\arguments{
\item{pAdjacency}{SEXP container for the pointer to the adjacency matrix}

\item{subsetIndices}{indices of the subset of the network to calculate
  the mean adjacency for.}
}
\value{
A List containing:
  \enumerate{
    \item{The mean absolute edge weight of the network subset.}
    \item{The weighted within-subset degree for each node.}
    \item{The maximum adjacency ratio for each node.}
    \item{The mean maximum adjacency ratio for the network subset.}
  }
}
\description{
Calculate Network Properties
}

