% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NetProps}
\alias{NetProps}
\title{Calculate Mean Adjacency and Intramodular Connectivity}
\usage{
NetProps(pAdjacency, subsetIndices)
}
\arguments{
\item{pAdjacency}{SEXP container for the pointer to the adjacency matrix.}

\item{subsetIndices}{indices of the network subset of interest.}
}
\value{
A List containing:
  \enumerate{
    \item{\emph{kIM}:}{The weighted within-subset degree for each node.}
    \item{\emph{mean.adj}:}{The mean absolute edge weight of the network subset.}
  }
}
\description{
Calculate Mean Adjacency and Intramodular Connectivity
}

