% Generated by roxygen2 (4.0.1): do not edit by hand
\name{netRep.core}
\alias{netRep.core}
\title{Assessing Network Replication}
\usage{
netRep.core(datSets = NULL, adjSets = NULL, nodeLabelSets, discovery, test,
  nPerm = 10000, buildNetFun, ignoreSets = NULL, includeSets = NULL,
  ignoreDiag = TRUE, verbose = TRUE, indent = 0)
}
\arguments{
\item{datSets}{\code{NULL}, or a list of \code{\link{big.matrix}}: one for
each dataset.}

\item{adjSets}{\code{NULL}, or a list of \code{\link{big.matrix}}: one for
each dataset, corresponding to the adjacency matrix of edge weights between
each pair of nodes in the network. Alternatively, if the list elements
are \code{NULL}, then the networks will be dynamically calculated using the
provided \code{buildNetFun} on the corresponding \code{datSet}.}

\item{nodeLabelSets}{a list, whose elements are \code{NULL} for each
\emph{test} dataset, and a vector for each \emph{discovery} dataset
assigning each node to a sub-network/cluster/module/component.}

\item{discovery}{a numeric vector indicating which elements of
\code{datSets} and/or \code{adjSets} are to be treated as the
\emph{discovery} datasets.}

\item{test}{a numeric vector indicating which elements of \code{datSets}
and/or \code{adjSets} are to be treated as the \emph{test} datasets.}

\item{nPerm}{number of permutations to use.}

\item{buildNetFun}{A function for constructing a network from the
\code{datSets}. This is used only where a pre-constructed network is not
provided in \code{adjSets}.}

\item{ignoreSets}{An optional list, where the elments for each
\emph{discovery} data set are vectors specifying which sub-networks to
skip.}

\item{includeSets}{An optional list, where the elments for each
\emph{discovery} data set are vectors specifying which sub-networks to
include.}

\item{ignoreDiag}{logical; ignore the diagonal of the adjacency matrix when
assessing network subset preservation. Defaults to \code{TRUE}.}

\item{verbose}{logical; print output while the function is running
(\code{TRUE} by default).}

\item{indent}{numeric; a positive value indicating the indent level to start
  the output at. Defaults to 0. Each indent level adds two spaces to the
  start of each line of output.}
}
\description{
This function provides the main functionality, but requires well formed
input. See details for instructions on usage. A wrapper function has been
provided for usability purposes, see \code{\link{netRep}} for details.
}
\details{
Any function argument ending in \code{"Sets"} expects a list, where each
 element corresponds to a dataset. \code{adjSets} should contain a list of
 adjacency matrices, one for each dataset. \code{datSets} should contain a
 list of \code{big.matrix} objects, which hold the underlying data the
 corresponding lement of the \code{adjSets} list (for example, a matrix of
 gene expression data). Here, we expect the rows of each element of
 \code{datSets} to correspond to a node in the network.

 Providing \code{datSets} is optional, if not provided, only statistics
 computable on the adjacency matrices will be calculated. If the underlying
 data is provided for any dataset, the corresponding element \code{adjSet}
 can be left as \code{NULL}, and the adjacency matrix will be constructed
 using the provided \code{buildNetFun} on the corresponding element of
 \code{datSet}.

 The elements of \code{nodeLabelSets} should be \code{NULL}, if its
 corresponding dataset is not a \emph{"discovery"} dataset, or contain a
 vector assigning each node in the network to a cluster/module/subnetwork
 etc. \code{netRep} will test each of these network subsets in the

 The order of the datasets should match across all \code{"Sets"}.

 If \code{datSets} is \code{NULL}, then this function assumes a full list of
 \code{adjSets} has been provided for each dataset of interest.

 \code{ignoreSets} and \code{includeSets} should not overlap.
}

