% Generated by roxygen2 (4.0.1): do not edit by hand
\name{NetStats}
\alias{NetStats}
\title{Calculate Network Statistics}
\usage{
NetStats(pAdjD, discIndices, pAdjT, testIndices)
}
\arguments{
\item{pAdjD,pAdjT}{SEXP containers for the pointers to the
\code{\link[bigmemory]{big.matrix}} objects holding the \emph{discovery}
and \emph{test} networks respectively.}

\item{discIndices,testIndices}{indices of the network subset of interest in
  the \emph{discovery} and \emph{test} networks respectively.}
}
\value{
A vector containing:
  \enumerate{
    \item{\emph{cor.adj}:}{
      The correlation between the edge weights for both networks.
    }
  }
}
\description{
For some statistics it does not make sense to calculate the necessary
components in advance due to large memory overhead, or logic that doesn't
separate nicely. This function deals with those statistics.
}

