% Generated by roxygen2 (4.0.2): do not edit by hand
\name{OrderNetwork}
\alias{OrderNetwork}
\title{Order nodes and clusters in a network}
\usage{
OrderNetwork(edge.matrix, node.labels = NULL, cluster.summaries = NULL)
}
\arguments{
\item{edge.matrix}{Matrix of edge weights}

\item{node.labels}{Optional vector of cluster assignments for each node.}

\item{cluster.summaries}{Optional data.frame of summary measurements to
compare clusters with.}
}
\value{
a vector of ordered nodes.
}
\description{
Nodes are ordered by degree, and clusters are ordered by similarity if some
underlying summary profile is provided in \code{cluster.summaries}
}

