% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parProgress}
\alias{parProgress}
\alias{setupParProgressLogs}
\alias{updateParProgress}
\title{Log Progress to File for Parallel Processes}
\usage{
setupParProgressLogs(chunk)

updateParProgress(pb, i)
}
\arguments{
\item{chunk}{The chunk of indices for the current parallel instance of the
\code{foreach} loop.}

\item{pb}{an object of class "\code{txtProgressBar}"}

\item{i}{new value for the progress bar.}
}
\value{
an object of class "\code{txtProgressBar}"
}
\description{
Update progress of a parallel \code{\link[foreach]{foreach}} process to
files.
}
\details{
When running code in parallel using \code{foreach}, new R
  instances will be spawned, one for each core as setup by the parallel
  backend. As a result no output will be sent back to the main R process
  until the \code{foreach} loop has finished running.
}
\seealso{
\code{\link[utils]{txtProgressBar}}
}

