% Generated by roxygen2 (4.0.2): do not edit by hand
\name{permutation}
\alias{perm.test}
\alias{permutation}
\alias{permuted}
\alias{requiredPerms}
\title{Permutation Test}
\usage{
perm.test(permuted, observed, subsetSize, totalSize, order = TRUE,
  alternative = "greater")

requiredPerms(alpha, alternative = "greater")
}
\arguments{
\item{permuted}{vector of values making up the empirical distribution.}

\item{observed}{the observed value of the test statistic.}

\item{subsetSize}{the size of the network subset the null distribution is
drawn for.}

\item{totalSize}{the size of the whole network}

\item{order}{logical; does the order of nodes in the permutation affect the
value of the test statistic?}

\item{alternative}{a character string specifying the alternative hypothesis,
 must be one of "greater" (default), "less", or "two.sided".
 You can specify just the initial letter.}

\item{alpha}{desired significance threshold.}
}
\value{
The minimum number of permutations required to detect any significant
 associations at the provided \code{alpha}. The minimum p-value will always
 be smaller than \code{alpha}.
}
\description{
Evaluates the statistical significance of a test statistic given a vector
 of "nulls": values for that test statistic drawn from random sampling.

\code{requiredPerms}:  how many permutations do I need to be able to detect
 significance at a given threshold \code{alpha}?
}
\details{
Calculates exact p-values for permutation tests when permutations are
 randomly drawn with replacement using the \code{\link[statmod]{permp}}
 function in the \code{\link{statmod}} package.
}
\references{
\enumerate{
    \item{
      Phipson, B. & Smyth, G. K. \emph{Permutation P-values should never be
      zero: calculating exact P-values when permutations are randomly drawn.}
      Stat. Appl. Genet. Mol. Biol. \strong{9}, Article39 (2010).
    }
  }
}

