% Generated by roxygen2 (4.0.1): do not edit by hand
\name{permutation}
\alias{permutation}
\alias{permuted}
\alias{pperm}
\alias{qperm}
\alias{requiredPower}
\alias{rperm}
\title{Distribution from Permutations}
\usage{
pperm(permuted, q, lower.tail = TRUE, log.p = FALSE)

qperm(permuted, p, log.p = FALSE)

rperm(permuted, n)

requiredPower(alpha)
}
\arguments{
\item{permuted}{vector of values making up the empirical distribution.}

\item{q}{vector of quantiles.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length is
taken to be the number required.}

\item{log.p}{logicial; if TRUE, probabilities p are given as log(p)}

\item{lower.tail}{logical; if TRUE (default), probabilities are
   \eqn{P[x \le x]} otherwise \eqn{P[X > x]}.}

\item{alpha}{desired significance threshold.}
}
\value{
The minimum number of permutations required to detect any significant
 associations at the provided \code{alpha}.
}
\description{
Distribution function, quantile function, and random generation
 for an empirical distribution provided by the user in the \code{permuted}
 argument. These distributions are typically drawn through permutation
 testing of some statistic.

\code{requiredPower}: If the tail approximation is not being used, how many
 permutations do I need to be able to detect significance at a given
 threshold \code{alpha}?
}
\details{
P-values are calculated by \code{pperm} using proportions on the provided
 distribution (\emph{permuted}). An observation, \code{q}'s, p-value
 is the proportion of \emph{permuted} whom are more extreme than \code{q}.
 This is accurate strictly when:
 \enumerate{
   \item Each permutation is independent.
   \item Permutations are sampled without replacement.
 }
 While these assumptions may not necessarily always be true, for our purposes
 they are a reasonable approximation. If \code{q} is more extreme than the
 \code{permuted}, the tail approximation method from \emph{(2)} can be used
 to estimate the p-value of \code{q} by setting \code{tail.approx} to
 \code{TRUE}.

 It is rarely possible to give an accurate value for \code{qperm} since there
 is no theoretical distribution to draw from. Instead the closest possible
 observation from the data is returned.
}
\note{
It is possible for the quantile \code{p} to fall exactly halfway
 between two observations from the \emph{permuted} distribution. In this case
 both data points are returned and a warning is generated. It is up to the
 user to choose which observation to take: the conservative approach is to
 choose the first observation if \code{p} < 0.5, or the second observation if
 \code{p} > 0.5.
}
\examples{
# A contrived example. For large n these results will be the same as qnorm,
 # rnorm, pnorm.
 normData <- rnorm(n=10000)
 pperm(normData, -1.644854) # should be approximately 0.05
 qperm(normData, 0.95) # should be approximately 1.644854
 rperm(normData, 100) # should be similar to rnorm(100)
}
\references{
\enumerate{
    \item{
      Phipson, B. & Smyth, G. K. \emph{Permutation P-values should never be
      zero: calculating exact P-values when permutations are randomly drawn.}
      Stat. Appl. Genet. Mol. Biol. \strong{9}, Article39 (2010).
    }
    \item{
      Knijnenburg, T. A., Wessels, L. F. A., Reinders, M. J. T. & Shmulevich,
      I. \emph{Fewer permutations, more accurate P-values}. Bioinformatics
      \strong{25}, i161-8 (2009).
    }
  }
}

