% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-defs.R
\name{plot_params}
\alias{plot_params}
\title{Template parameters}
\arguments{
\item{orderNodesBy}{\code{NULL} (default), \code{NA}, or a vector of dataset
names or indices. Controls how nodes are ordered on the plot (see details).}

\item{orderSamplesBy}{\code{NULL} (default), \code{NA}, or a vector 
containing a single dataset name or index. Controls how samples are ordered 
on the plot (see details).}

\item{plotNodeNames}{logical; controls whether the node names are 
rendered on the bottom axis.}

\item{plotSampleNames}{logical; controls whether the sample names are 
rendered on the left axis.}

\item{plotModuleNames}{logical; controls whether module names are rendered.
The default is for module names to be rendered when multiple \code{modules} 
are drawn.}

\item{drawBorders}{logical; if \code{TRUE}, borders are drawn around the 
connectivity, module membership, and module summary bar plots.}

\item{border.width}{line width for borders.}

\item{naxt.line}{the number of lines into the bottom margin at which the node
names will be drawn.}

\item{saxt.line}{the number of lines into the left margin at which the sample
names will be drawn.}

\item{maxt.line}{the number of lines into the bottom margin at which the 
module names will be drawn.}

\item{legend.tick.size}{size of the ticks on each axis legend relative to the
size of the correlation, edge weights, and data matrix heatmaps.}

\item{laxt.line}{the distance from the legend to render the legend axis 
labels, as multiple of \code{legend.tick.size}.}

\item{cex.axis}{relative size of the node and sample names.}

\item{cex.lab}{relative size of the module names and legend titles.}

\item{cex.main}{relative size of the plot titles.}
}
\description{
Template parameters to be imported into other function documentation. This 
is not intended to be a stand-alone help file.
}

