% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotBar}
\alias{plotBar}
\title{Custom bar plot function}
\usage{
plotBar(heights, heights.lim, mas, cols, bar.width = 1, drawBorder = FALSE)
}
\arguments{
\item{heights}{heights of the bars.}

\item{heights.lim}{limits of the height axis.}

\item{mas}{ordered subset of the moduleAssignments vector}

\item{cols}{colors of each bar.}

\item{bar.width}{value between 0 and 1 controlling the proportion of space
taken by each bar.}

\item{drawBorder}{logical; if \code{TRUE} a border is drawn around each bar.}
}
\description{
Plot bars around 0
}

