% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotCoexpression}
\alias{plotCoexpression}
\title{Coexpression plot}
\usage{
plotCoexpression(coexpression, module.labels = NULL, heatmap.gradient)
}
\arguments{
\item{module.labels}{an optional named vector assigning each node to a module.}

\item{heatmap.gradient}{A vector of colors to use for (or interpolate over)
to render the coexpression.}

\item{coexpresion}{a square matrix containing the pairwise gene coexpression.}
}
\description{
Plot a triangle of the coexpression
}

