% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotExpression}
\alias{plotExpression}
\title{Plot a heatmap of the gene expression}
\usage{
plotExpression(gene.expr, heatmap.gradient, expression.range,
  is.relative = TRUE, cex.title = 1.4, main = "Gene Expression",
  xlab = "genes", ylab = "samples", cex.lab = 1)
}
\arguments{
\item{gene.expr}{matrix of gene expression data. Expects columns to be the
genes.}

\item{heatmap.gradient}{A vector of colors to use for (or interpolate over)
to render the gene expression.}

\item{expression.range}{range of values the gene expression can take.}

\item{is.relative}{logical; is the expression relative, i.e. centred around 0?
This affects automated color gradient choice.}

\item{cex.title}{cex for the title text}

\item{main}{title for the plot}

\item{xlab}{label for the x axis}

\item{ylab}{label for the y axis}

\item{cex.lab}{cex for the axis label}
}
\description{
Plot a heatmap of the gene expression
}

