% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-module.R
\name{plotModule}
\alias{plotModule}
\title{Plot the topology of a network module}
\usage{
plotModule(data, correlation, network, moduleAssignments = NULL,
  modules = NULL, backgroundLabel = "0", discovery = NULL, test = NULL,
  nCores = NULL, verbose = TRUE, orderSamplesBy = "test",
  orderNodesBy = "discovery", orderModules = TRUE, plotNodeNames = TRUE,
  plotSampleNames = TRUE, plotModuleNames, main = "Module Topology",
  drawBorders = FALSE, border.width = 2, gaxt.line = -0.5,
  saxt.line = -0.5, maxt.line, legend.tick.size = 0.04, laxt.line = 2.5,
  cex.axis = 0.8, cex.lab = 1, cex.main = 1.2)
}
\arguments{
\item{data}{a list of numeric matrices. Each entry of the list corresponds to
a dataset and contains the data used to infer the interaction network
between variables (e.g. genes). Expects matrix columns to correspond to
variables and matrix rows to correspond to samples.}

\item{correlation}{a list of matrices. Each entry of the list corresponds to a 
dataset and contains an \eqn{n * n} matrix of the correlation between 
each pair of variables in the dataset.}

\item{network}{a list of matrices. Each entry of the list corresponds to a 
dataset and contains an \eqn{n * n} matrix of the network edge weights 
between each pair of variables in the dataset.}

\item{moduleAssignments}{a vector containing the module each variable belongs
to in the discovery dataset. If there are multiple discovery datasets 
then this argument should be a list of such vectors.}

\item{modules}{a list of vectors, one for each \code{discovery} dataset, 
of modules to perform the analysis on. The default is to analyse all modules
with the exception of those specified in \code{backgroundLabel}.}

\item{backgroundLabel}{a single label that nodes that do not belong to any
module are assigned. The default is "0".}

\item{discovery}{a vector of names or indices denoting the discovery dataset(s).}

\item{test}{a list of vectors of names or indices denoting the test datasets
for each \code{discovery} dataset. Alternatively can be provided as vector
if the test datasets are the same for all 'discovery' datasets (e.g. for 
performing a pairwise comparison).}

\item{nCores}{number of cores to parallelise the calculation of network 
properties over. Ignored if the user has already registered a parallel 
backend.If \code{NULL} (default) the maximum number of cores on the machine
will be used.}

\item{verbose}{logical; should progress be reported? Default is \code{TRUE}.}

\item{orderSamplesBy}{one of "discovery", "test" or "none". Controls how 
samples are ordered on the plot (see details).}

\item{orderNodesBy}{one of "discovery", "test" or "none". Controls how nodes
are ordered on the plot (see details).}

\item{orderModules}{logical; if \code{TRUE} modules ordered by clustering 
their summary vectors. If \code{FALSE} modules are returned in the order
provided.}

\item{plotNodeNames}{logical; controls whether the node names are 
rendered on the bottom axis.}

\item{plotSampleNames}{logical; controls whether the sample names are 
rendered on the left axis.}

\item{plotModuleNames}{logical; controls whether module names are rendered.
The default is for module names to be rendered when multiple \code{modules} 
are drawn.}

\item{main}{title for the plot.}

\item{drawBorders}{logical; if \code{TRUE}, borders are drawn around the 
connectivity, module membership, and module summary bar plots.}

\item{border.width}{line width for borders.}

\item{gaxt.line}{the number of lines into the bottom margin at which the node
names will be drawn.}

\item{saxt.line}{the number of lines into the left margin at which the sample
names will be drawn.}

\item{maxt.line}{the number of lines into the bottom margin at which the 
module names will be drawn.}

\item{legend.tick.size}{size of the ticks on each axis legend relative to the
size of the correlation, edge weights, and data matrix heatmaps.}

\item{laxt.line}{the distance from the legend to render the legend axis 
labels, as multiple of \code{legend.tick.size}.}

\item{cex.axis}{relative size of the node and sample names.}

\item{cex.lab}{relative size of the module names and legend titles.}

\item{cex.main}{relative size of the plot titles.}
}
\description{
Plot the correlation structure, network edges, (normalised) connectivity, 
module membership, underlying data, and module summary vector of one or
more network modules.

Individual components of the module plot can be plotted using 
\code{\link{plotCorrelation}}, \code{\link{plotNetwork}}, 
\code{\link{plotDegree}}, \code{\link{plotContribution}}, 
\code{\link{plotData}}, and \code{\link{plotSummary}}.
}
\details{
\subsection{Input data structure:}{
  This function allows for input data formatted in a number of ways. Where 
  there are multiple datasets of interest (e.g. multiple tissues, locations, 
  or a discovery dataset and an independent test dataset) the arguments 
  \code{data}, \code{correlation}, and \code{network} should be
  \code{\link[=list]{lists}} where each element contains the matrix data for 
  each respective dataset. Alternatively, if only one dataset is of interest, 
  the \code{data}, \code{correlation}, and \code{network} arguments
  will also each accept a 'matrix' object.
  
  Similarly, the \code{moduleAssignments} argument expects a list of named
  vectors, which denote the module each variable belongs to in the discovery
  dataset. If module discovery has only been performed in one dataset, then 
  the \code{moduleAssignments} argument will also accept a named vector.
  
  The \code{discovery} arguments specifies which dataset the \code{modules} 
  of interest were discovered in, and the \code{test} argument specifies 
  which dataset to calculate the network properties in. These arguments are
  ignored if data is provided for only one dataset.
}
\subsection{'bigMatrix' vs. 'matrix' input data:}{
  Although the function expects \code{\link[=bigMatrix-class]{bigMatrix}} 
  data, regular 'matrix' objects are also accepted. In this case, the 
  'matrix' data is temporarily converted to 'bigMatrix' by the function. This
  conversion process involves writing out each matrix as a binary file on 
  disk, which can take a long time for large datasets. It is strongly 
  recommended for the user to store their data as 'bigMatrix' objects, as the
  \link{modulePreservation} function, \link[=plotModule]{plotting} 
  \link[=plotTopology]{functions}, \link[=nodeOrder]{node} and 
  \link[=sampleOrder]{sample} ordering functions also expect 'bigMatrix'
  objects. Further, 'bigMatrix' objects have a number of benefits, including 
  instantaneous load time from any future R session, and parallel access from
  mutliple independent R sessions. Methods are provided for 
  \link[=bigMatrix-get]{converting to, loading in}, and 
  \link[=bigMatrix-out]{writing out} 'bigMatrix' objects.
}
\subsection{Node, sample, and module ordering:}{
  By default, nodes are ordered in decreasing order of \emph{weighted degree}
  in the \code{discovery} dataset (see \code{\link{nodeOrder}}). 
  This facilitates the visual comparison of modules across datasets, as the 
  node ordering will be preserved. Missing nodes are colored in grey. If
  \code{orderNodesBy} is "test" nodes will instead be ordered by 
  weighted degree in the \code{test} dataset. If "none" nodes are 
  drawn in the order they are provided in the drawn dataset.
  
  When multiple modules are specified, modules are ordered by the similarity
  of their summary vectors in the drawn dataset. To disable this behaviour, 
  set \code{orderModules} to \code{FALSE}.
  
  Sample ordering only applies to \code{plotData} and 
  \code{plotModuleSummary}. By default, samples are ordered in descending
  order of the module summary vector in the drawn dataset for the left-most 
  module appearing on the plot (see \code{\link{sampleOrder}}.
}
\subsection{Normalised degree:}{
  The weighted degree is normalised by the maximum connectivity in
  any given module when rendered on the bar plot. This facilitates visual 
  comparison of multiple modules with differing sizes or densities.
}
\subsection{Plot customisation:}{
  Although reasonable default values for most parameters have been provided,
  the rendering of axes and titles may need adjusting depending on the size
  of the plot window. The parameters \code{gaxt.line}, \code{saxt.line}, 
  \code{maxt.line}, and \code{laxt.line} control the distance from each plot
  window that the node labels, sample labels, module labels, and legend 
  labels are rendered. 
  
  \code{legend.tick.size} controls the length of the 
  axis ticks on each of the legends relative to the correlation, edge weight,
  and data matrix heatmap plot windows. 
  
  \code{cex.main} controls the relative text size of the plot title
  (specified by the \code{main} argument). \code{cex.axis} controls the
  relative text size of the node and sample labels. \code{cex.lab} controls
  the relative text size of the bar plot axis labels, module labels, and the
  legend titles.
  
  The rendering of node, sample, and module names can be disabled by setting
  \code{plotNodeNames}, \code{plotSampleNames}, and \code{plotModuleNames} to
  \code{FALSE}.
  
  The \code{drawBorders} argument controls whether borders are drawn around
  the connectivity, module membership, or module summary bar plots.
}
}
\examples{
\dontrun{
## Create some example data
geA <- matrix(rnorm(50*100), ncol=100) # gene expression
colnames(geA) <- paste0("Gene_", 1:100)
rownames(geA) <- paste0("CohortA_", 1:50)
coexpA <- cor(geA) # correlation
adjA <- abs(coexpA)^5 # network
moduleAssignments <- sample(1:7, size=100, replace=TRUE)
names(moduleAssignments) <- paste0("Gene_", 1:100)

# Create bigMatrix objects for each matrix.
geA <- as.bigMatrix(geA, "geA_bm")
coexpA <- as.bigMatrix(coexpA, "coexpA_bm")
adjA <- as.bigMatrix(adjA, "adjA_bm")

## Example 1: Plot Module 2 in cohort A.
plotModule(geA, coexpA, adjA, moduleAssignments, modules="2")

## Example 2: Plot an arbitrary set of genes in cohort A
plotModule(geA[,1:10], coexpA[1:10, 1:10], adjA[1:10, 1:10])

## Example 3: Plot the topology of two adipose tissue modules in the liver
## tissue data 

geAdipose <- matrix(rnorm(50*100), ncol=100) # gene expression
colnames(geAdipose) <- paste0("Gene_", 1:100)
rownames(geAdipose) <- paste0("Sample_", 1:50)
coexpAdipose <- cor(geAdipose) # correlation
adjAdipose <- abs(coexpAdipose)^5 # network
adiposeModules <- sample(0:7, size=100, replace=TRUE)
names(adiposeModules) <- paste0("Gene_", 1:100)

geLiver <- matrix(rnorm(50*100), ncol=100) # gene expression
colnames(geLiver) <- paste0("Gene_", 1:100)
rownames(geLiver) <- paste0("Sample_", 1:50)
coexpLiver <- cor(geLiver) # correlation
adjLiver <- abs(coexpLiver)^6 # network
liverModules <- sample(0:12, size=100, replace=TRUE)
names(liverModules) <- paste0("Gene_", 1:100)

geHeart <- matrix(rnorm(50*100), ncol=100) # gene expression
colnames(geHeart) <- paste0("Gene_", 1:100)
rownames(geHeart) <- paste0("Sample_", 1:50)
coexpHeart <- cor(geHeart) # correlation
adjHeart <- abs(coexpHeart)^4 # network
heartModules <- sample(0:5, size=100, replace=TRUE)
names(heartModules) <- paste0("Gene_", 1:100)

# Store each input type as a list, where each element corresponds
# to one of the datasets
data <- list(
  adipose=as.bigMatrix(geAdipose, "geAdipose_bm"),
  liver=as.bigMatrix(geLiver, "geLiver_bm"),  
  heart=as.bigMatrix(geHeart, "geHeart_bm") 
)
correlation <- list(
  adipose=as.bigMatrix(coexpAdipose, "coexpAdipose_bm"),
  liver=as.bigMatrix(coexpLiver, "coexpLiver_bm"),  
  heart=as.bigMatrix(coexpHeart, "coexpHeart_bm") 
)
network <- list(
  adipose=as.bigMatrix(adjAdipose, "adjAdipose_bm"),
  liver=as.bigMatrix(adjLiver, "adjLiver_bm"),  
  heart=as.bigMatrix(adjHeart, "adjHeart_bm") 
)
moduleAssignments <- list(
  adipose=adiposeModules, liver=liverModules, heart=heartModules
)

# Show the plot
plotModule(
  data, correlation, network, moduleAssignments,
  modules=c("3", "7"), discovery="adipose", test="liver"
)

# clean up bigMatrix files from examples
unlink("*_bm*")
}

}
\seealso{
\code{\link{plotCorrelation}}, 
\code{\link{plotNetwork}},
\code{\link{plotDegree}},
\code{\link{plotContribution}},
\code{\link{plotData}}, and
\code{\link{plotSummary}}.
}

