% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotMultiBar}
\alias{plotMultiBar}
\title{Plot multiple horizontal bar plots}
\usage{
plotMultiBar(lengths, lengths.lim, cols, bar.width = 1, drawBorder = FALSE,
  main = "")
}
\arguments{
\item{lengths}{a matrix whose columns contain the lengths of each bar for
the given property (e.g. each column should be a module, or phenotype).}

\item{lengths.lim}{a list of limits for the lengths axes.}

\item{cols}{a matrix of colors for each bar.}

\item{bar.width}{value between 0 and 1 controlling the proportion of space
taken by each bar.}

\item{drawBorder}{logical; if \code{TRUE} a border is drawn around each bar.}

\item{main}{title for the plot}
}
\description{
Plot multiple horizontal bar plots
}

