% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream-analysis.R
\name{plotProps}
\alias{plotProps}
\title{Get the network properties and order for a plot}
\usage{
plotProps(scaledData, correlation, network, moduleAssignments, modules, di, ti,
  orderNodesBy, orderSamplesBy, orderModules, datasetNames, nDatasets, dryRun,
  verbose)
}
\arguments{
\item{scaledData}{scaled data returned by \code{'processInput'}.}

\item{correlation}{list returned by \code{'processInput'}.}

\item{network}{list returned by \code{'processInput'}.}

\item{moduleAssignments}{list returned by \code{'processInput'}.}

\item{modules}{vector of modules to show on the plot.}

\item{di}{name of the discovery dataset.}

\item{ti}{name of the test dataset.}

\item{orderNodesBy}{vector returned by \code{'processInput'}.}

\item{orderSamplesBy}{vector returned by \code{'processInput'}.}

\item{orderModules}{vector returned by \code{'checkPlotArgs'}.}

\item{datasetNames}{vector returned by \code{'processInput'}.}

\item{nDatasets}{vector returned by \code{'processInput'}.}

\item{dryRun}{logical; are we just doing a dry run of the plot?}

\item{verbose}{logical; turn on verbose printing.}
}
\description{
Get the network properties and order for a plot
}

