% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotSummaryExpression}
\alias{plotSummaryExpression}
\title{Plot the Summary Expression Profile of a Module}
\usage{
plotSummaryExpression(gene.expr, heatmap.gradient, expression.range,
  is.relative = TRUE, cex.axis = 1, cex.title = 1.4,
  main = "Summary\\nExpression", ylab = "samples", cex.lab = cex.axis,
  missing.inds = NULL, new.samples = 0)
}
\arguments{
\item{heatmap.gradient}{A vector of colors to use for (or interpolate over)
to render the coexpression.}

\item{is.relative}{logical; is the expression relative, i.e. centred around 0?
This affects automated color gradient choice.}

\item{cex.axis}{cex for the axis text}

\item{cex.title}{cex for the title text}

\item{main}{title for the plot}

\item{cex.lab}{cex for the axis label}

\item{missing.inds}{locations for the missing samples}

\item{new.samples}{the number of new samples in the dataset. Assumes they are
  placed at the end of the expression matrix, so draws a line to indicate
  where these fall.}

\item{coexpresion}{a square matrix containing the pairwise gene coexpression.}

\item{module.labels}{an optional named vector assigning each node to a module.}
}
\description{
Vertical barplot.
}

