% Generated by roxygen2 (4.0.2): do not edit by hand
\name{preservationPlot}
\alias{preservationPlot}
\title{Generate preservation plot components}
\usage{
preservationPlot(gene.expr, coexpression, adjacency, moduleLabels, module,
  moduleGenes, sampleOrder, sampleNames, includeNew = TRUE,
  is.relative = TRUE, cex.axis = 0.8, cex.title = 0.9, cex.lab = 0.6)
}
\arguments{
\item{gene.expr}{gene expression matrix. Expects rows to be samples, and
columns to be genes.}

\item{coexpression}{matrix of pairwise gene coexpression.}

\item{adjacency}{matrix of pairwise gene adjacencies.}

\item{moduleLabels}{vector of labels assigning each gene to a module.}

\item{module}{module to plot preservation for.}

\item{moduleGenes}{alternative to \code{module}: specify module genes /
indices instead to preserve order (see details). Can either be a character
vector of gene names, or a vector of indices for the corresponding
\code{moduleLabels}.}

\item{sampleOrder}{(optional) order for the samples. If omitted, samples are
ordered based on their summary expression profile. This is useful for
preserving sample order across datasets (e.g. tissues) (see details).}

\item{sampleNames}{(optional) set of sample names for the dataset. If
omitted, and \code{sampleOrder} has been provided, and is a set of names,
this will default to the rownames of the \code{gene.expr}.}

\item{includeNew}{logical; In the case \code{sampleOrder} and
\code{sampleNames} have been provided, should the samples present in the
data provided, but not in \code{sampleOrder} be included in the plots?}

\item{is.relative}{logical; is the gene expression relative to some other
measure? If \code{TRUE}, a divergent palette is selected when plotting the
expression, otherwise a sequential palette is chosen.}

\item{cex.axis}{\code{cex} for the axis tick labels}

\item{cex.title}{\code{cex} for the titles}

\item{cex.lab}{\code{cex} for the axis labels}
}
\value{
Outputs files for the following:
\itemize{
\item coexpression heatmap
\item coexpression legend
\item intramodular connectivity
\item intramodular module membership
\item expression heatmap
\item summary expression profile
}
}
\description{
Generate preservation plot components
}
\details{
All matrix arguments will take \code{matrix}, \code{big.matrix}, or
\code{big.matrix} descriptors.

If \code{module} is specified, genes will be ordered by connectivity.
Alternatively, if \code{moduleGenes} is given, nodes will remain in the same
order as specified. \code{\link{orderNetwork}} will be useful to determine
the correct order. If \code{sampleOrder} is provided, samples in the gene
expression, and summary expression plots will be ordered as described.
Missing samples will be shown in grey. Samples not present in the
\code{sampleOrder} (e.g. not present in the discovery dataset) will be shown
at the bottom, ordered by their summary expression.

Where genes are missing, these will be represented with grey squares in all
plots.
}

