% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-user-input.R
\name{processInput}
\alias{processInput}
\title{Process, check, and format input}
\usage{
processInput(discovery, test, network, correlation, data, moduleAssignments,
  modules, backgroundLabel, verbose, funcType, orderNodesBy = NA,
  orderSamplesBy = NA, orderModules = NULL)
}
\arguments{
\item{discovery}{user input for the 'discovery' argument.}

\item{test}{user input for the 'test' argument.}

\item{network}{user input for the 'network' argument.}

\item{correlation}{user input for the 'correlation' argument.}

\item{data}{user input for the 'data' argument.}

\item{moduleAssignments}{user input for the 'moduleAssignments' argument.}

\item{modules}{user input for the 'modules' argument.}

\item{backgroundLabel}{user input for the 'backgroundLabel' argument.}

\item{verbose}{logical; should progress be reported? Default is \code{TRUE}.}

\item{funcType}{one of "preservation", "properties" or "plot".}

\item{orderNodesBy}{user input for the 'orderNodesBy' argument in the 
plotting functions.}

\item{orderSamplesBy}{user input for the 'orderSamplesBy' argument in the 
plotting functions.}

\item{orderModules}{user input for the 'orderModules' argument in the 
plotting functions.}
}
\value{
a list of containing the formatted user input
}
\description{
Checks user input for consistency, errors, and unifies the data structures.
}
\seealso{
\code{\link{modulePreservation}}
\code{\link{plotModule}}, and
\code{\link{plotTopology}}
}
\keyword{internal}

