% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RangeSubset}
\alias{BigRange}
\alias{RangeSubset}
\alias{range-cpp}
\title{Get the range of a big.matrix}
\usage{
RangeSubset(pDat, subsetIndices)

BigRange(pDat)
}
\arguments{
\item{pDat}{SEXP container for the pointer to the data matrix to be scaled.}

\item{subsetIndices}{indices of the network subset of interest in 
\code{pDat}.}
}
\description{
\code{RangeSubset}: get the range of values in the column-subset of a 
 big.matrix.

\code{BigRange}: get the range of values in a big.matrix
}

