% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigMatrix-functions.R
\name{rangeBigMatrix}
\alias{rangeBigMatrix}
\title{Get the range of a bigMatrix or its subset}
\usage{
rangeBigMatrix(x, subsetIndices)
}
\arguments{
\item{x}{a bigMatrix}

\item{subsetIndices}{an optional vector to subset the matrix by}
}
\description{
Note subsetting only applies to columns: this is only meant for use with the 
'data' matrices.
}

