% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scaleBigMatrix}
\alias{scaleBigMatrix}
\title{Scale and Center the rows of a 'big.matrix'}
\usage{
scaleBigMatrix(x, backingfile = NULL, backingpath = NULL,
  descriptorfile = NULL)
}
\arguments{
\item{x}{the \code{\link[bigmemory]{big.matrix}} to scale.}

\item{backingfile,backingpath,descriptorfile}{backingfile information for the
new scaled big.matrix to be stored in, see
\code{\link[bigmemory]{big.matrix}}.}
}
\value{
A new \code{\link[bigmemory]{big.matrix}} containing the scaled and centered
 rows of \code{x}.
}
\description{
Create a new scaled \code{\link[bigmemory]{big.matrix}}.
}
\note{
Unlike \code{\link[base]{scale}} this function applies the scale function to
 the rows of the matrix instead of the columns.

 This function has only been implemented for
 \code{\link[bigmemory]{big.matrix}} objects of \code{type = "double"}.

 To apply the scale function in place without creating a new results matrix,
 simply call \code{Scale(x@address, x@address)}.
}
\seealso{
\code{\link[base]{scale}} \code{\link[bigmemory]{big.matrix}}
}

