% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigMatrix-functions.R
\name{scaleBigMatrix}
\alias{scaleBigMatrix}
\title{Scale and Center the rows of a 'big.matrix'}
\usage{
scaleBigMatrix(x, tmp.dir)
}
\arguments{
\item{x}{the bigMatrix to scale.}

\item{tmp.dir}{temporary directory to store the backingfiles for the 
bigMatrix.}
}
\value{
an object of type \code{bigMatrix}.
}
\description{
Create a new \code{\link[=bigMatrix-class]{bigMatrix}} containing scaled gene
expression. Genes are expected to be columns.
}
\note{
The caller is responsible for cleaning up the temporary directory.
}
\seealso{
\code{\link[base]{scale}} \code{\link[bigmemory]{big.matrix}}
}

