% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setupParallel}
\alias{setupParallel}
\title{Set up a parallel backend}
\usage{
setupParallel(nCores = NULL, verbose)
}
\arguments{
\item{nCores}{number of cores to use.}
}
\description{
Set up a backend with the requested number of cores, or use existing backend
if the user has set one up already.
}

