% Generated by roxygen2 (4.0.1): do not edit by hand
\name{subsetProps}
\alias{subsetProps}
\title{Network subset topology}
\usage{
subsetProps(adj, adjInd, dat = NULL, scaled = NULL, datInd = NULL)
}
\arguments{
\item{adj}{Adjacency matrix for the network.}

\item{adjInd}{Indices of the network subset in \code{adj}.}

\item{dat}{(Optional) Underlying data for the network.}

\item{scaled}{(Optional) a row scaled \code{big.matrix} of \code{dat}.}

\item{datInd}{(Optional) Indices of the network subset in \code{dat}.}
}
\value{
A list of topological properties for the given network subset
}
\description{
Calculate the topological characteristics of a network subset.
}
\seealso{
\code{\link[=subsetTestStats]{Between-network statistics}}
}

