% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp-wrappers.R
\name{wrappers}
\alias{corStats}
\alias{dataProps}
\alias{moduleProps}
\alias{netProps}
\alias{wrappers}
\title{Fast wrapper functions for Rcpp functions}
\usage{
corStats(discCor, discIndices, testCor, testIndices)

netProps(adj, moduleIndices)

dataProps(sdat, moduleIndices)

moduleProps(adj, moduleIndices, sdat)
}
\arguments{
\item{discCor}{a \code{\link{bigMatrix}} containing the
correlation structure among variables in the \emph{discovery} dataset.}

\item{discIndices}{indices corresponding to the network module in 
\code{discCor}.}

\item{testCor}{a \code{\link{bigMatrix}} containing the
correlation structure among variables in the \emph{test} dataset.}

\item{testIndices}{indices corresponding to the network module in
\code{testCor}.}

\item{adj}{a \code{\link{bigMatrix}} containing an
adjacency matrix representation of the network (i.e. an \eqn{n * n} matrix 
containing the edge weights between each pair of variables).}

\item{moduleIndices}{indices for the network module of interest in the
supplied data matrix.}

\item{sdat}{a \code{\link{bigMatrix}} containing scaled 
numeric data. Assumes columns correspond to variables (e.g. genes, 
microbial operational taxonomic unit) and rows correspond to samples.}
}
\value{
\code{corStats:} a list containing:
 \enumerate{
   \item{\emph{cor.discovery}:}{
      A flattened vector of the module's correlation structure in the 
      \emph{discovery} dataset.
   } \item{\emph{cor.test}:}{
      A flattened vector of the module's correlation structure in the 
      \emph{test} dataset.
   } \item{\emph{corDensity}:}{
     The mean sign-aware correlation structure density of the network module.
   }
 }

\code{netProps:} a list of containing:  
 \enumerate{
   \item{\emph{weightedDegree}:}{
     The weighted degree (within the module) for each node.
   }
   \item{\emph{averageEdgeWeight}:}{
     The mean absolute edge weight of the network subset.
   }
 }

\code{dataProps:} a list containing:
  \enumerate{
   \item{\emph{moduleSummary}:}{The module summary vector.}
   \item{\emph{nodeContribution}:}{
     The contribution of each node to the module's summary profile , i.e. 
     the correlation between each variable and the module summary vector.
   }
   \item{\emph{moduleCoherence}:}{
     The proportion of module variance explained by the module summary vector. 
   }
  }
}
\description{
Wrapper functions for interfacing with the Rcpp implementations of 
\code{\link{CorStats}}, \code{\link{NetProps}}, and \code{\link{DataProps}}.
They provide a nicer interface for input handling and format the output in
an R-friendly way.

combines \code{dataProps} and \code{netProps}
}
\references{
\enumerate{
    \item{
      Langfelder, P., Luo, R., Oldham, M. C. & Horvath, S. \emph{Is my
      network module preserved and reproducible?} PLoS Comput. Biol. 
      \strong{7}, e1001057 (2011). 
    }
 }
}

