% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wrappers}
\alias{adjProps}
\alias{coexpStats}
\alias{dataProps}
\alias{moduleProps}
\alias{wrappers}
\title{Fast wrapper functions for Rcpp functions}
\usage{
coexpStats(discCoexp, discIndices, testCoexp, testIndices)

adjProps(adj, moduleIndices)

dataProps(sge, moduleIndices)

moduleProps(adj, moduleIndices, sge)
}
\arguments{
\item{discCoexp}{a \code{\link{bigMatrix}} containing the pairwise
coexpression for the \emph{discovery} dataset.}

\item{discIndices}{indices corresponding to the network module in the matrix
of pairwise-coexpression in the \emph{discovery} dataset.}

\item{testCoexp}{a \code{\link{bigMatrix}} containing the pairwise
coexpression for the \emph{test} dataset.}

\item{testIndices}{indices corresponding to the network module, or a random
  sample of the same size, in the matrix of pairwise-coexpression in the
  \emph{test} dataset.}

\item{adj}{a \code{\link{bigMatrix}} containing the pairwise gene
 adjacencies.}

\item{moduleIndices}{indices for the network module of interest in the
  supplied data matrix.}

\item{sge}{a \code{\link{bigMatrix}} containing scaled gene expression data.
  Assumes columns are genes/probes, rows are samples, and that probes have
  been scaled (This can be done using \code{\link{scaleBigMatrix}}).}
}
\value{
\code{coexpStats:} a list containing the \emph{cor.coexp} and
 \emph{mean.coexp} statistics for the specified module across datasets.

\code{netProps:} a list of network properties calculated from an adjacency
 matrix. These properties can either be scalers (summarising the whole
 network subset), or vectors (characterising some property for each node in
 the network subset).

\code{dataProps:} a list of properties quantifying the relationship between
  a network subset and the underlying data the adjacency matrix was
  calculated from. These properties can either be scalers (summarising the
  whole network subset), or vectors (characterising some property for each
  node in the network subset).
}
\description{
Wrapper functions for interfacing with the Rcpp implementations of
\code{\link{CoexpStats}}, \code{\link{AdjProps}}, and \code{\link{DataProps}}.
They provide a nicer interface for input handling and format the output in
an R-friendly way.

combines \code{dataProps} and \code{adjProps}
}

